Harden PentestGPT integration: extract activity, strict findings, network/CI/docs
Phase 1 (extract dedicated Pentest activity, no behavior change):
- New TemporalPentestActivities in moonmind/workflows/temporal/activities/
  pentest_activities.py holding all Pentest-specific scope/policy/provider/
  materialization/workload/publish/findings/heartbeat/cleanup logic.
- TemporalAgentRuntimeActivities.security_pentest_execute and the trusted
  internal entrypoint are now thin delegates; the security.pentest.execute
  binding is unchanged. Patched module-level symbols are referenced via the
  activity_runtime module so existing monkeypatch tests still apply.
- Tests prove the handler attr resolves to the public untrusted entrypoint,
  the trusted entrypoint is never registered, and untrusted dispatch can never
  self-authorize an inline approved_scope.

Phase 6 (strict findings validation):
- strictly_normalize_pentest_finding_set validates each record against
  PentestFinding, quarantines malformed records (no arbitrary dicts), counts
  only valid findings, and sets normalization_status so reports never imply
  "no vulnerabilities" on upstream/provider/runner/normalizer failure.
- Activity-side loader and the in-runner normalize_findings.py both apply this;
  the runner wrapper emits an honest, status-aware summary/report.

Phase 4 (network semantics, Option A): rename safe profile network_policy
restricted_egress -> bridge_approved_lab and document that it is Docker bridge
for approved lab targets, not enforced egress; external targets stay disabled.

Phase 5 (runner CI): add `moonmind-pentestgpt-run --check-upstream` upstream
CLI compatibility check (distinct exit codes), wire it into docker-publish.yml,
surface the published digest, and fail if the default runner tag is unpublished.

Phases 7/8: discovery tests assert the tool is hidden when disabled, VPN profile
is hidden unless enabled, full_authorized/evidence/time-budget are narrowed from
settings, and no raw shell/docker/image/secret inputs are exposed.

Phases 9/10: staging rollout checklist and rollout gates in PentestOperations.md.

Phases 2/3: opt-in, local/staging-only integration scaffolding for the real
Docker launcher and provider profile manager (excluded from required CI).
