+++ docker/pentestgpt-runner/moonmind-pentestgpt-run
@@ -23,7 +23,7 @@ PY
 
 usage() {
   cat <<'USAGE'
-moonmind-pentestgpt-run [--target TARGET] [--instruction-file PATH] [--non-interactive] [--no-telemetry] [--version] [--self-test]
+moonmind-pentestgpt-run [--target TARGET] [--instruction-file PATH] [--non-interactive] [--no-telemetry] [--version] [--self-test] [--check-upstream]
 USAGE
 }
 
@@ -43,6 +43,32 @@ while [[ $# -gt 0 ]]; do
     --self-test)
       SELF_TEST=1
       shift
+      ;;
+    --check-upstream)
+      # Verify the installed upstream PentestGPT CLI contract. Distinct exit
+      # codes let CI classify upstream-CLI breakage separately from wrapper or
+      # configuration failures.
+      if ! command -v pentestgpt >/dev/null 2>&1; then
+        echo "upstream pentestgpt executable not found" >&2
+        exit 30
+      fi
+      if ! pentestgpt --help >/dev/null 2>&1; then
+        echo "upstream pentestgpt --help failed to parse" >&2
+        exit 31
+      fi
+      # Confirm telemetry stays disabled by default in the image.
+      if [[ "${LANGFUSE_ENABLED:-false}" != "false" ]]; then
+        echo "LANGFUSE_ENABLED must default to false in the runner image" >&2
+        exit 32
+      fi
+      # Confirm wrapper redaction works on a sample secret-like string.
+      REDACTED_SAMPLE="$(printf 'api_key=ghp_examplesecrettoken1234567890' | redact)"
+      if [[ "$REDACTED_SAMPLE" == *"ghp_examplesecrettoken1234567890"* ]]; then
+        echo "wrapper redaction failed on sample secret" >&2
+        exit 33
+      fi
+      echo "upstream-check ok"
+      exit 0
       ;;
     --target)
       TARGET="${2:-}"
@@ -188,11 +214,60 @@ print(json.dumps({
 PY
 )"
 
+NORMALIZER_RC=0
 python /usr/local/lib/moonmind-pentestgpt/normalize_findings.py \
   --target "$TARGET" \
   --mode "$MODE" \
   --input "$RAW_EVIDENCE_FILE" \
-  --output "$STRUCTURED_FILE" 2> >(redact > "$STDERR_FILE") || EXIT_CODE=$?
+  --output "$STRUCTURED_FILE" 2> >(redact >> "$STDERR_FILE") || NORMALIZER_RC=$?
+if [[ "$NORMALIZER_RC" -ne 0 ]]; then
+  echo "normalize_findings.py failed with code $NORMALIZER_RC" | redact >> "$STDERR_FILE"
+  EXIT_CODE="$NORMALIZER_RC"
+fi
+
+# Derive an honest, status-aware summary/report from the normalized findings so
+# the human-facing report never implies "no vulnerabilities" on failure.
+read -r NORM_STATUS FINDINGS_COUNT HIGH_CRIT QUARANTINED < <(python - "$STRUCTURED_FILE" "$NORMALIZER_RC" <<'PY'
+import json
+import sys
+from pathlib import Path
+
+path = Path(sys.argv[1])
+normalizer_rc = sys.argv[2]
+try:
+    data = json.loads(path.read_text(encoding="utf-8"))
+except Exception:
+    data = None
+if not isinstance(data, dict):
+    print("normalizer_error 0 0 0")
+else:
+    summary = data.get("summary") or {}
+    status = data.get("normalization_status") or "no_machine_readable_findings"
+    if normalizer_rc != "0":
+        status = "normalizer_error"
+    print(
+        status,
+        int(summary.get("findings_count") or 0),
+        int(summary.get("high_or_critical_count") or 0),
+        int(data.get("quarantined_findings_count") or 0),
+    )
+PY
+)
+
+case "$NORM_STATUS" in
+  ok)
+    HEADLINE="$FINDINGS_COUNT finding(s) reported ($HIGH_CRIT high/critical)." ;;
+  completed_no_findings)
+    HEADLINE="Completed with no findings. No vulnerabilities were reported for this run." ;;
+  upstream_failed)
+    HEADLINE="The upstream PentestGPT tool failed; results are incomplete. This is NOT a clean result." ;;
+  no_valid_findings)
+    HEADLINE="All candidate findings failed validation and were quarantined; results are inconclusive." ;;
+  no_machine_readable_findings)
+    HEADLINE="The tool produced no machine-readable findings; review raw evidence. This is NOT a clean result." ;;
+  *)
+    HEADLINE="Normalization failed ($NORM_STATUS); review raw evidence and diagnostics. This is NOT a clean result." ;;
+esac
 
 cat > "$SUMMARY_FILE" <<MD
 # Pentest Summary
@@ -201,7 +276,11 @@ Target: $TARGET
 
 Mode: $MODE
 
-No confirmed findings were emitted by the wrapper. Review raw evidence and runtime diagnostics for provider/tool failures or incomplete execution.
+Status: $NORM_STATUS
+
+$HEADLINE
+
+Quarantined (malformed) records: $QUARANTINED. Review raw evidence and runtime diagnostics for provider/tool failures or incomplete execution.
 MD
 
 cat > "$REPORT_FILE" <<MD
@@ -213,9 +292,17 @@ cat > "$REPORT_FILE" <<MD
 - Operation mode: $MODE
 - Instruction file: $INSTRUCTION_FILE
 
+## Result Status
+
+- Normalization status: \`$NORM_STATUS\`
+- Findings: $FINDINGS_COUNT ($HIGH_CRIT high/critical)
+- Quarantined malformed records: $QUARANTINED
+
+$HEADLINE
+
 ## Findings
 
-The runner produced the normalized findings artifact at \`$STRUCTURED_FILE\`.
+The runner produced the schema-valid normalized findings artifact at \`$STRUCTURED_FILE\`.
 
 ## Evidence
