+++ docker/pentestgpt-runner/normalize_findings.py
@@ -1,5 +1,19 @@
 #!/usr/bin/env python3
-"""Normalize runner evidence into MoonMind's PentestFindingSet envelope."""
+"""Normalize runner evidence into MoonMind's PentestFindingSet envelope.
+
+Runs *inside* the curated runner image, so it cannot import MoonMind. It
+mirrors the strict ``PentestFinding`` / ``PentestFindingSet`` contract:
+
+* each raw finding is validated against the schema rules and **quarantined**
+  (never passed through as an arbitrary dict) if it fails;
+* the compact summary only counts valid findings;
+* ``normalization_status`` distinguishes a genuine clean result from the
+  failure/empty modes, so the human-facing report cannot imply
+  "no vulnerabilities" when the upstream tool or normalizer failed.
+
+The MoonMind activity boundary re-validates this output with the real Pydantic
+models, so this script is a best-effort in-runner gate, not the sole guarantee.
+"""
 
 from __future__ import annotations
 
@@ -7,6 +21,130 @@ import argparse
 from datetime import datetime, timezone
 import json
 from pathlib import Path
+from typing import Any
+
+_VALID_SEVERITIES = {"info", "low", "medium", "high", "critical"}
+_VALID_CONFIDENCES = {"suspected", "supported", "confirmed"}
+_STATUSES = {
+    "ok",
+    "completed_no_findings",
+    "no_machine_readable_findings",
+    "upstream_failed",
+    "normalizer_error",
+    "no_valid_findings",
+}
+
+
+def _normalize_confidence(value: Any) -> str:
+    normalized = str(value or "").strip().lower()
+    if normalized in {"supported", "confirmed"}:
+        return normalized
+    return "suspected"
+
+
+def _validate_finding(record: Any, *, target: str) -> tuple[dict[str, Any] | None, str | None]:
+    """Return ``(finding, None)`` if valid, else ``(None, reason)``."""
+
+    if not isinstance(record, dict):
+        return None, "record must be an object"
+    finding_target = str(record.get("target") or "").strip() or target
+    finding_id = str(record.get("finding_id") or "").strip()
+    title = str(record.get("title") or "").strip()
+    summary = str(record.get("summary") or "").strip()
+    severity = str(record.get("severity") or "").strip().lower()
+    if not finding_id:
+        return None, "finding_id is required"
+    if not title:
+        return None, "title is required"
+    if not finding_target:
+        return None, "target is required"
+    if severity not in _VALID_SEVERITIES:
+        return None, f"invalid severity: {severity or '(empty)'}"
+    normalized = {
+        "finding_id": finding_id,
+        "title": title,
+        "severity": severity,
+        "confidence": _normalize_confidence(record.get("confidence")),
+        "target": finding_target,
+        "summary": summary or "No finding summary provided.",
+    }
+    for optional in ("category", "impact", "remediation", "reproduction_artifact_ref"):
+        if record.get(optional):
+            normalized[optional] = str(record[optional])
+    for seq in ("evidence_artifact_refs", "related_ids"):
+        value = record.get(seq)
+        if isinstance(value, list):
+            normalized[seq] = [str(item) for item in value if str(item).strip()]
+    return normalized, None
+
+
+def normalize_evidence(
+    raw: Any,
+    *,
+    target: str,
+    mode: str,
+    evidence_ref: str,
+) -> dict[str, Any]:
+    """Normalize raw runner evidence into the schema-valid envelope."""
+
+    if not isinstance(raw, dict):
+        raw = {}
+    upstream_exit_code = raw.get("upstream_exit_code")
+    try:
+        upstream_exit_code = (
+            int(upstream_exit_code) if upstream_exit_code is not None else None
+        )
+    except (TypeError, ValueError):
+        upstream_exit_code = None
+
+    raw_findings = raw.get("findings")
+    findings_key_present = "findings" in raw
+    findings_is_list = isinstance(raw_findings, list)
+
+    valid: list[dict[str, Any]] = []
+    quarantined: list[dict[str, Any]] = []
+    if findings_is_list:
+        for index, record in enumerate(raw_findings):
+            finding, reason = _validate_finding(record, target=target)
+            if finding is not None:
+                valid.append(finding)
+            else:
+                quarantined.append({"index": index, "reason": reason})
+
+    if upstream_exit_code not in (None, 0):
+        status = "upstream_failed"
+    elif not findings_key_present or not findings_is_list:
+        status = "no_machine_readable_findings"
+    elif valid:
+        status = "ok"
+    elif quarantined:
+        status = "no_valid_findings"
+    else:
+        status = "completed_no_findings"
+
+    return {
+        "tool_name": "security.pentest.run",
+        "tool_version": "1.0.0",
+        "target": target,
+        "scope_artifact_ref": str(raw.get("scope_artifact_ref") or ""),
+        "operation_mode": mode,
+        "runner_profile_id": str(raw.get("runner_profile_id") or "pentestgpt-safe"),
+        "execution_profile_ref": raw.get("execution_profile_ref") or None,
+        "produced_at": datetime.now(timezone.utc).isoformat().replace("+00:00", "Z"),
+        "findings": valid,
+        "summary": {
+            "findings_count": len(valid),
+            "confirmed_findings_count": sum(
+                1 for item in valid if item["confidence"] == "confirmed"
+            ),
+            "high_or_critical_count": sum(
+                1 for item in valid if item["severity"] in {"high", "critical"}
+            ),
+        },
+        "normalization_status": status,
+        "quarantined_findings_count": len(quarantined),
+        "evidence_refs": [evidence_ref],
+    }
 
 
 def main() -> int:
@@ -24,34 +162,13 @@ def main() -> int:
         raw = json.loads(input_path.read_text(encoding="utf-8"))
     except Exception:
         raw = {}
-    findings = raw.get("findings")
-    if not isinstance(findings, list):
-        findings = []
-    findings = [item for item in findings if isinstance(item, dict)]
 
-    normalized = {
-        "tool_name": "security.pentest.run",
-        "tool_version": "1.0.0",
-        "target": args.target,
-        "scope_artifact_ref": "",
-        "operation_mode": args.mode,
-        "runner_profile_id": "pentestgpt-safe",
-        "execution_profile_ref": None,
-        "produced_at": datetime.now(timezone.utc).isoformat().replace("+00:00", "Z"),
-        "findings": findings,
-        "summary": {
-            "findings_count": len(findings),
-            "confirmed_findings_count": sum(
-                1 for item in findings if item.get("confidence") == "confirmed"
-            ),
-            "high_or_critical_count": sum(
-                1
-                for item in findings
-                if item.get("severity") in {"high", "critical"}
-            ),
-        },
-        "evidence_refs": [evidence_ref],
-    }
+    normalized = normalize_evidence(
+        raw,
+        target=args.target,
+        mode=args.mode,
+        evidence_ref=evidence_ref,
+    )
     output_path.parent.mkdir(parents=True, exist_ok=True)
     output_path.write_text(
         json.dumps(normalized, sort_keys=True, indent=2) + "\n",
