+++ docs/Security/PentestOperations.md
@@ -106,6 +106,33 @@ The Docker publish workflow builds and publishes this image alongside the main M
 
 The runner entrypoint is `/usr/local/bin/moonmind-pentestgpt-run`. It runs non-interactively, disables Langfuse telemetry by default, captures redacted stdout/stderr, writes MoonMind report artifacts, and normalizes findings JSON.
 
+CI smoke tests the wrapper with `--version` and `--self-test`, and validates the
+installed upstream PentestGPT CLI contract with `--check-upstream` (the
+`pentestgpt` executable exists, `pentestgpt --help` parses, telemetry defaults
+to off, and wrapper redaction works). Upstream-CLI breakage uses distinct exit
+codes (30–33) so it is classified separately from wrapper/configuration
+failures.
+
+## Network Enforcement
+
+**`pentestgpt-safe` runs on the Docker `bridge` network. This is NOT a
+network-enforced "restricted egress" boundary.** The runner can, at the OS
+level, reach any network the Docker host can reach. The first production-safe
+posture therefore relies on **approved-scope validation**, not on network
+enforcement:
+
+- `MOONMIND_PENTEST_ALLOW_EXTERNAL_TARGETS=false` keeps targets confined to
+  lab/internal-authorized classes recorded in the approved scope artifact.
+- The `pentestgpt-safe` profile is intended for **approved lab targets only**.
+- Operator docs, launch-plan metadata, runner profile names, and UI copy must
+  describe this profile as Docker `bridge` for approved lab targets — never as
+  an enforced egress firewall.
+
+Real restricted egress (a dedicated Docker network, egress proxy, or firewall
+sidecar that can only reach approved lab/provider endpoints) is **future work**
+and gated behind a separate security review. Until that exists, do not enable
+external targets and do not represent the safe profile as network-enforced.
+
 ## Task Inputs
 
 Task-level inputs remain narrow:
@@ -179,3 +206,56 @@ MOONMIND_PENTEST_TELEMETRY_ENABLED=false
 
 4. Submit a `recon_only` task first.
 5. Confirm report artifacts and cleanup metadata before trying `validate_hypothesis`.
+
+## VPN / Lab Profile (Deferred)
+
+VPN support is **disabled by default and not exposed** in production:
+
+- `MOONMIND_PENTEST_ALLOW_VPN_LAB_PROFILE=false` keeps `pentestgpt-vpn-lab`
+  hidden from tool discovery and rejected by deployment policy.
+- A curated VPN profile (explicit `NET_ADMIN`, `/dev/net/tun`, a named VPN
+  state volume, a required `network_attachment_ref`, no arbitrary host mounts,
+  and a separate security review) is **future work**.
+- Until that profile is implemented and reviewed, leave VPN disabled.
+
+## Staging Rollout Checklist
+
+Run this checklist when bringing the integration up in a staging environment:
+
+1. Build or pull the approved runner image.
+2. Prefer a digest-pinned `MOONMIND_PENTEST_RUNNER_IMAGE` for staging/production.
+3. Configure `MOONMIND_WORKFLOW_DOCKER_MODE=profiles`.
+4. Run an `agent_runtime` worker on `mm.activity.agent_runtime`.
+5. Confirm Docker proxy access (`SYSTEM_DOCKER_HOST` / `DOCKER_HOST`,
+   `MOONMIND_DOCKER_PROXY_NETWORK`).
+6. Confirm `agent_workspaces` volume access
+   (`MOONMIND_AGENT_WORKSPACES_VOLUME_NAME` if overridden).
+7. Configure the Temporal artifact backend (`TEMPORAL_ARTIFACT_BACKEND` and
+   related S3/local settings).
+8. Configure the provider profile and secret refs
+   (`MOONMIND_PENTEST_DEFAULT_PROVIDER_PROFILE`,
+   `MOONMIND_PENTEST_PROVIDER_PROFILES_FILE`,
+   `MOONMIND_PENTEST_ENABLED_PROVIDER_PROFILES`,
+   `PENTESTGPT_ANTHROPIC_SECRET_ID` / `PENTESTGPT_OPENROUTER_SECRET_ID`).
+9. Create a lab-only approved scope artifact (`target_class=lab`).
+10. Set `MOONMIND_PENTEST_ENABLED=true`.
+11. Submit `pentest-recon` with `operation_mode=recon_only`.
+12. Verify report bundle refs (`report.primary`, `report.summary`,
+    `report.structured`, `report.evidence`).
+13. Verify no secret leaks in logs, artifacts, workflow history, or UI previews.
+14. Verify cleanup metadata and provider lease release.
+15. Disable again (`MOONMIND_PENTEST_ENABLED=false`) unless the environment is
+    approved for continued operation.
+
+## Rollout Gates
+
+Promote capability in this order; each step is gated on the previous one
+passing and on security review for the escalations:
+
+1. Merge hardening behind `MOONMIND_PENTEST_ENABLED=false`.
+2. Run local/dev lab tests with the real runner image.
+3. Run staging tests with the real provider profile manager.
+4. Enable `recon_only` for internal lab scopes.
+5. Enable `validate_hypothesis` for internal authorized scopes.
+6. Keep `full_authorized`, VPN profiles, and external targets disabled until a
+   separate security review signs off.
