+++ docs/Steps/PentestTool.md
@@ -61,6 +61,38 @@ The implementation includes:
 
 The default deployment posture remains conservative: `MOONMIND_PENTEST_ENABLED=false`, approved scope is required, telemetry is disabled, only the safe runner profile is allowlisted, VPN/lab profiles are hidden, and `full_authorized` is not allowlisted unless an operator explicitly enables it.
 
+#### Production hardening status
+
+Hardening completed in this change:
+
+- The Pentest-specific activity logic is extracted into a dedicated
+  `TemporalPentestActivities` class
+  (`moonmind/workflows/temporal/activities/pentest_activities.py`);
+  `TemporalAgentRuntimeActivities.security_pentest_execute` is now a thin
+  delegate, so the `security.pentest.execute` binding is unchanged.
+- Normalized findings are strictly validated against `PentestFinding` /
+  `PentestFindingSet`; malformed records are quarantined (never passed through
+  as arbitrary dicts) and a `normalization_status` distinguishes a genuine
+  clean result from upstream/provider/runner/normalizer failures, so reports
+  never imply "no vulnerabilities" on failure.
+- Network semantics are reconciled to **Option A**: the safe profile runs on
+  Docker `bridge` (`network_policy=bridge_approved_lab`) and relies on
+  approved-scope validation, not enforced egress. External targets stay
+  disabled. See `docs/Security/PentestOperations.md` → *Network Enforcement*.
+- Runner CI validates the upstream PentestGPT CLI contract via
+  `moonmind-pentestgpt-run --check-upstream`, surfaces the published image
+  digest, and fails if the default runner tag is not published.
+- VPN/lab profile remains deferred (hidden/disabled).
+- Staging rollout checklist and rollout gates are documented in
+  `docs/Security/PentestOperations.md`.
+
+Validation that requires real infrastructure (run in dev/staging, not in
+required CI — see `tests/integration/temporal/test_pentest_docker_e2e.py`):
+
+- Phase 2: real Docker workload-launcher E2E (opt-in via `MOONMIND_PENTEST_E2E=1`).
+- Phase 3: real provider profile manager E2E (opt-in via
+  `MOONMIND_PENTEST_PROVIDER_E2E=1`).
+
 ---
 
 ## 2. Summary
@@ -626,7 +658,7 @@ resource_profile:
   cpu: "4"
   memory: "8g"
   shm_size: "1g"
-network_policy: restricted_egress
+network_policy: bridge_approved_lab
 device_policy: none
 timeout_seconds: 28800
 cleanup:
