+++ moonmind/integrations/pentest/__init__.py
@@ -11,6 +11,7 @@ from .models import (
     PentestFindingSet,
     PentestFindingSummary,
     PentestLiveLogEvent,
+    PentestNormalizationStatus,
     PentestPublicationArtifact,
     PentestPublicationDiagnostics,
     PentestPublicationResult,
@@ -47,6 +48,7 @@ from .models import (
     get_pentest_provider_profile,
     materialize_pentest_provider_profile,
     normalize_pentest_confidence,
+    strictly_normalize_pentest_finding_set,
     pentest_orphan_cleanup_candidate_matches,
     pentest_cleanup_selector,
     pentest_container_name,
@@ -69,6 +71,7 @@ __all__ = [
     "PentestFinding",
     "PentestFindingSet",
     "PentestFindingSummary",
+    "PentestNormalizationStatus",
     "PentestInstructionBundle",
     "PentestLaunchPlan",
     "PentestLaunchPolicyError",
@@ -104,6 +107,7 @@ __all__ = [
     "get_pentest_runner_profile",
     "materialize_pentest_provider_profile",
     "normalize_pentest_confidence",
+    "strictly_normalize_pentest_finding_set",
     "pentest_orphan_cleanup_candidate_matches",
     "pentest_cleanup_selector",
     "pentest_container_name",
