+++ moonmind/integrations/pentest/models.py
@@ -15,6 +15,7 @@ from pydantic import (
     BaseModel,
     ConfigDict,
     Field,
+    ValidationError,
     ValidationInfo,
     field_validator,
     model_validator,
@@ -37,9 +38,26 @@ PentestScopeAction = Literal[
     "exploit_validation",
 ]
 PentestNetworkAttachmentType = Literal["none", "vpn", "lab_network"]
-PentestNetworkPolicy = Literal["restricted_egress", "approved_lab_egress"]
+# Network policy is descriptive metadata, NOT an enforced firewall. The safe
+# profile runs on Docker ``bridge`` and relies on approved-scope validation, so
+# it is named ``bridge_approved_lab`` to avoid implying enforced egress control.
+# ``approved_lab_egress`` is reserved for the (deferred) curated VPN/lab profile.
+PentestNetworkPolicy = Literal["bridge_approved_lab", "approved_lab_egress"]
 PentestFindingSeverity = Literal["info", "low", "medium", "high", "critical"]
 PentestFindingConfidence = Literal["suspected", "supported", "confirmed"]
+# Normalization status distinguishes "no vulnerabilities" from failure modes so
+# the human-facing report never implies a clean result when the provider/tool or
+# the normalizer failed, or when the upstream produced nothing machine-readable.
+PentestNormalizationStatus = Literal[
+    "ok",
+    "completed_no_findings",
+    "no_machine_readable_findings",
+    "upstream_failed",
+    "provider_failed",
+    "runner_failed",
+    "normalizer_error",
+    "no_valid_findings",
+]
 PentestArtifactClass = Literal["input", "runtime", "output", "evidence"]
 PentestFailureKind = Literal[
     "image_pull",
@@ -653,6 +671,123 @@ class PentestFindingSet(PentestBaseModel):
     produced_at: datetime
     findings: tuple[PentestFinding, ...] = ()
     summary: PentestFindingSummary
+    # Distinguishes a genuine clean result from failure/empty modes so the
+    # human-facing report cannot imply "no vulnerabilities" on failure.
+    normalization_status: PentestNormalizationStatus = "ok"
+    # Count of records that failed strict validation and were quarantined out of
+    # ``findings``. The quarantined records themselves are recorded in restricted
+    # diagnostics/evidence, never inlined into the compact result payload.
+    quarantined_findings_count: int = Field(default=0, ge=0)
+
+    @property
+    def implies_no_vulnerabilities(self) -> bool:
+        """True only when a clean result can be honestly reported."""
+
+        return (
+            self.normalization_status in {"ok", "completed_no_findings"}
+            and self.summary.findings_count == 0
+            and self.quarantined_findings_count == 0
+        )
+
+
+def strictly_normalize_pentest_finding_set(
+    payload: Any,
+    *,
+    target: str,
+    operation_mode: str,
+    runner_profile_id: str,
+    scope_artifact_ref: str = "",
+    execution_profile_ref: str | None = None,
+    produced_at: datetime | None = None,
+    upstream_exit_code: int | None = None,
+    status_hint: PentestNormalizationStatus | None = None,
+) -> tuple[PentestFindingSet, list[dict[str, Any]]]:
+    """Validate runner findings into a schema-valid :class:`PentestFindingSet`.
+
+    Each raw finding record is validated against :class:`PentestFinding`. Records
+    that fail validation are **quarantined** (never passed through as arbitrary
+    dicts) and returned separately so callers can record them in restricted
+    diagnostics/evidence rather than the compact result payload.
+
+    The returned set's ``normalization_status`` distinguishes a genuine clean
+    result from failure/empty modes, so the human-facing report cannot imply
+    "no vulnerabilities" when the upstream/tool/normalizer failed.
+
+    Returns ``(finding_set, quarantined)`` where ``quarantined`` is a list of
+    ``{"index", "reason", "record"}`` dicts (record values redacted).
+    """
+
+    produced_at = produced_at or datetime.now(UTC)
+    raw_findings: Any = None
+    if isinstance(payload, dict):
+        raw_findings = payload.get("findings")
+
+    valid: list[PentestFinding] = []
+    quarantined: list[dict[str, Any]] = []
+    findings_key_present = isinstance(payload, dict) and "findings" in payload
+    findings_is_list = isinstance(raw_findings, list)
+
+    if findings_is_list:
+        for index, record in enumerate(raw_findings):
+            # Backfill the run's target onto findings that omit it — a common
+            # upstream shape when the run has a single target. This is genuine
+            # normalization, not passing arbitrary dicts through.
+            if isinstance(record, dict) and not record.get("target") and target:
+                record = {**record, "target": target}
+            try:
+                valid.append(PentestFinding.model_validate(record))
+            except ValidationError as exc:
+                reasons = ", ".join(
+                    f"{'.'.join(str(p) for p in err.get('loc', ()))}: {err.get('type')}"
+                    for err in exc.errors()
+                )
+                quarantined.append(
+                    {
+                        "index": index,
+                        "reason": redact_pentest_human_text(reasons or "invalid finding"),
+                        "record": redact_pentest_diagnostic_value(record),
+                    }
+                )
+
+    if status_hint is not None:
+        status: PentestNormalizationStatus = status_hint
+    elif upstream_exit_code not in (None, 0):
+        status = "upstream_failed"
+    elif payload is None:
+        status = "normalizer_error"
+    elif not findings_key_present or not findings_is_list:
+        status = "no_machine_readable_findings"
+    elif valid:
+        status = "ok"
+    elif quarantined:
+        status = "no_valid_findings"
+    else:
+        status = "completed_no_findings"
+
+    summary = PentestFindingSummary(
+        findings_count=len(valid),
+        confirmed_findings_count=sum(
+            1 for finding in valid if finding.confidence == "confirmed"
+        ),
+        high_or_critical_count=sum(
+            1 for finding in valid if finding.severity in {"high", "critical"}
+        ),
+    )
+
+    finding_set = PentestFindingSet(
+        target=target,
+        scope_artifact_ref=scope_artifact_ref or "",
+        operation_mode=operation_mode,  # type: ignore[arg-type]
+        runner_profile_id=runner_profile_id,
+        execution_profile_ref=execution_profile_ref,
+        produced_at=produced_at,
+        findings=tuple(valid),
+        summary=summary,
+        normalization_status=status,
+        quarantined_findings_count=len(quarantined),
+    )
+    return finding_set, quarantined
+
 
 class PentestLiveLogEvent(PentestBaseModel):
     """MoonMind-native Pentest Live Log projection event."""
@@ -1516,7 +1651,7 @@ def _pentest_runner_profiles() -> dict[str, PentestRunnerProfile]:
             optional_mounts=(),
             env_allowlist=_PENTEST_ENV_ALLOWLIST,
             resource_profile=_PENTEST_RESOURCE_PROFILE,
-            network_policy="restricted_egress",
+            network_policy="bridge_approved_lab",
             linux_capabilities=(),
             devices=(),
             timeout_seconds=28800,
