+++ moonmind/workflows/temporal/activity_runtime.py
@@ -61,6 +61,7 @@ from moonmind.integrations.pentest.models import (
     redact_pentest_diagnostic_value,
     redact_pentest_human_text,
     resolve_pentest_provider_profile,
+    strictly_normalize_pentest_finding_set,
 )
 from moonmind.jules.status import JulesStatusSnapshot, normalize_jules_status
 from moonmind.schemas.manifest_ingest_models import CompiledManifestPlanModel
@@ -5144,598 +5145,13 @@ class TemporalAgentRuntimeActivities:
             or TemporalPentestProviderLeaseManager(client_adapter)
         )
         self._supervision_tasks: set[asyncio.Task] = set()
-
-    @staticmethod
-    def _pentest_scope_artifact_id(scope_artifact_ref: object) -> str:
-        if isinstance(scope_artifact_ref, ArtifactRef):
-            artifact_id = str(scope_artifact_ref.artifact_id or "").strip()
-        elif isinstance(scope_artifact_ref, Mapping):
-            artifact_id = str(scope_artifact_ref.get("artifact_id") or "").strip()
-        else:
-            artifact_id = str(scope_artifact_ref or "").strip()
-        if not artifact_id:
-            raise PentestScopeValidationError(
-                error_code="INVALID_SCOPE",
-                reasons=("scope_artifact_ref_required",),
-                diagnostics={"scope_artifact_ref": None},
-            )
-        return artifact_id
-
-    async def _load_pentest_approved_scope(
-        self,
-        request_payload: Mapping[str, Any],
-    ) -> PentestApprovedScope:
-        if self._artifact_service is None:
-            raise PentestScopeValidationError(
-                error_code="INVALID_SCOPE",
-                reasons=("scope_artifact_service_unavailable",),
-                diagnostics={"scope_artifact_ref": request_payload.get("scope_artifact_ref")},
-            )
-        artifact_id = self._pentest_scope_artifact_id(
-            request_payload.get("scope_artifact_ref")
-        )
-        principal = str(
-            request_payload.get("principal_id")
-            or request_payload.get("agent_run_id")
-            or "workflow"
-        ).strip()
-        try:
-            _artifact, payload = await self._artifact_service.read(
-                artifact_id=artifact_id,
-                principal=principal,
-            )
-        except TemporalArtifactError as exc:
-            raise PentestScopeValidationError(
-                error_code="INVALID_SCOPE",
-                reasons=("scope_artifact_unreadable",),
-                diagnostics={"scope_artifact_ref": artifact_id},
-            ) from exc
-        try:
-            decoded = json.loads(payload.decode("utf-8"))
-        except (UnicodeDecodeError, json.JSONDecodeError) as exc:
-            raise PentestScopeValidationError(
-                error_code="INVALID_SCOPE",
-                reasons=("scope_artifact_malformed",),
-                diagnostics={"scope_artifact_ref": artifact_id},
-            ) from exc
-        try:
-            return PentestApprovedScope.model_validate(decoded)
-        except ValidationError as exc:
-            raise PentestScopeValidationError(
-                error_code="INVALID_SCOPE",
-                reasons=("scope_artifact_invalid",),
-                diagnostics={
-                    "scope_artifact_ref": artifact_id,
-                    "error_count": exc.error_count(),
-                },
-            ) from exc
-
-    @staticmethod
-    def _validate_pentest_request_policy(
-        request: PentestWorkloadRequest,
-        scope: PentestApprovedScope,
-    ) -> None:
-        pentest_settings = settings.pentest
-        allowed_profiles = set(pentest_settings.allowed_runner_profiles)
-        if (
-            pentest_settings.allow_vpn_lab_profile
-            and pentest_settings.vpn_lab_profile_id
-        ):
-            allowed_profiles.add(pentest_settings.vpn_lab_profile_id)
-        if request.runner_profile_id not in allowed_profiles:
-            raise PentestLaunchPolicyError(
-                "Pentest runner profile is not enabled by deployment policy",
-                reason="runner_profile_disabled",
-                diagnostics={
-                    "runner_profile_id": request.runner_profile_id,
-                    "allowed_runner_profiles": sorted(allowed_profiles),
-                },
-            )
-        if request.operation_mode not in set(pentest_settings.allowed_operation_modes):
-            raise PentestLaunchPolicyError(
-                "Pentest operation mode is not enabled by deployment policy",
-                reason="operation_mode_disabled",
-                diagnostics={
-                    "operation_mode": request.operation_mode,
-                    "allowed_operation_modes": list(
-                        pentest_settings.allowed_operation_modes
-                    ),
-                },
-            )
-        if request.evidence_level not in set(pentest_settings.allowed_evidence_levels):
-            raise PentestLaunchPolicyError(
-                "Pentest evidence level is not enabled by deployment policy",
-                reason="evidence_level_disabled",
-                diagnostics={
-                    "evidence_level": request.evidence_level,
-                    "allowed_evidence_levels": list(
-                        pentest_settings.allowed_evidence_levels
-                    ),
-                },
-            )
-        if request.time_budget_minutes > pentest_settings.max_time_budget_minutes:
-            raise PentestLaunchPolicyError(
-                "Pentest time budget exceeds deployment policy",
-                reason="time_budget_exceeded",
-                diagnostics={
-                    "time_budget_minutes": request.time_budget_minutes,
-                    "max_time_budget_minutes": pentest_settings.max_time_budget_minutes,
-                },
-            )
-        if (
-            scope.target_class == "external_authorized"
-            and not pentest_settings.allow_external_targets
-        ):
-            raise PentestScopeValidationError(
-                error_code="UNAPPROVED_TARGET",
-                reasons=("external_targets_disabled",),
-                diagnostics={
-                    "scope_id": scope.scope_id,
-                    "target": request.target,
-                    "target_class": scope.target_class,
-                },
-            )
-        if (
-            scope.target_class == "external_authorized"
-            and pentest_settings.require_manual_approval_for_external
-            and not scope.approval_recorded
-        ):
-            raise PentestScopeValidationError(
-                error_code="UNAPPROVED_TARGET",
-                reasons=("external_target_manual_approval_required",),
-                diagnostics={"scope_id": scope.scope_id, "target": request.target},
-            )
-        if (
-            request.operation_mode == "full_authorized"
-            and pentest_settings.require_manual_approval_for_full_authorized
-            and not scope.approval_recorded
-        ):
-            raise PentestScopeValidationError(
-                error_code="UNAPPROVED_TARGET",
-                reasons=("full_authorized_manual_approval_required",),
-                diagnostics={"scope_id": scope.scope_id, "target": request.target},
-            )
-
-    def _validate_pentest_runner_profiles_registered(self) -> None:
-        if self._workload_registry is None or not hasattr(
-            self._workload_registry, "get"
-        ):
-            return
-        pentest_settings = settings.pentest
-        configured_profiles = set(pentest_settings.allowed_runner_profiles)
-        configured_profiles.add(pentest_settings.default_runner_profile)
-        configured_profiles.add(pentest_settings.safe_profile_id)
-        if pentest_settings.allow_vpn_lab_profile:
-            configured_profiles.add(pentest_settings.vpn_lab_profile_id)
-        missing_profiles = sorted(
-            profile_id
-            for profile_id in configured_profiles
-            if profile_id and self._workload_registry.get(profile_id) is None
-        )
-        if missing_profiles:
-            raise PentestLaunchPolicyError(
-                "Configured Pentest runner profile is not present in workload registry",
-                reason="runner_profile_not_registered",
-                diagnostics={"runner_profile_ids": missing_profiles},
-            )
-        drifted_profiles = []
-        for profile_id in configured_profiles:
-            if not profile_id:
-                continue
-            profile = self._workload_registry.get(profile_id)
-            if profile is not None and profile.image != pentest_settings.runner_image:
-                drifted_profiles.append(profile_id)
-        drifted_profiles.sort()
-        if drifted_profiles:
-            raise PentestLaunchPolicyError(
-                "Configured Pentest runner profile image does not match approved runner image",
-                reason="runner_profile_image_drift",
-                diagnostics={
-                    "runner_profile_ids": drifted_profiles,
-                    "approved_runner_image": pentest_settings.runner_image,
-                },
-            )
-
-    @staticmethod
-    def _write_pentest_json(
-        path: str,
-        payload: Mapping[str, Any],
-    ) -> None:
-        redacted_payload = redact_sensitive_payload(dict(payload))
-        target = Path(path)
-        target.parent.mkdir(parents=True, exist_ok=True)
-        target.write_text(
-            json.dumps(redacted_payload, sort_keys=True, indent=2) + "\n",
-            encoding="utf-8",
+        # Pentest-specific activity logic lives in a dedicated module/class.
+        # Imported lazily to avoid an import cycle (that module imports this one).
+        from moonmind.workflows.temporal.activities.pentest_activities import (
+            TemporalPentestActivities,
         )
 
-    async def _resolve_pentest_secret_env(
-        self,
-        provider_materialization: Any,
-    ) -> tuple[dict[str, str], tuple[str, ...]]:
-        resolved: dict[str, str] = {}
-        missing: list[str] = []
-        for env_key, logical_ref in provider_materialization.secret_env.items():
-            ambient = str(os.environ.get(env_key, "") or "").strip()
-            if ambient:
-                resolved[env_key] = ambient
-                continue
-
-            secret_ref_config = provider_materialization.secret_refs.get(logical_ref)
-            candidate_refs: list[str | Mapping[str, Any]] = []
-            if isinstance(secret_ref_config, str):
-                candidate_refs.append(secret_ref_config)
-            elif isinstance(secret_ref_config, Mapping):
-                candidate_refs.append(secret_ref_config)
-                for key in ("secret_ref", "secretRef", "secret_id", "secretId"):
-                    value = str(secret_ref_config.get(key) or "").strip()
-                    if value:
-                        candidate_refs.append(value)
-            for candidate in candidate_refs:
-                if isinstance(candidate, str) and not re.match(
-                    r"^[a-z][a-z0-9+.-]*://", candidate
-                ):
-                    continue
-                try:
-                    resolved_value = await resolve_managed_api_key_reference(
-                        candidate,
-                        field_name=f"pentest.{env_key}",
-                    )
-                except Exception:
-                    logger.warning(
-                        "Failed to resolve PentestGPT configured secret reference",
-                    )
-                    continue
-                if str(resolved_value or "").strip():
-                    resolved[env_key] = str(resolved_value)
-                    break
-            if env_key not in resolved:
-                missing.append(env_key)
-        return resolved, tuple(missing)
-
-    def _materialize_pentest_input_files(
-        self,
-        *,
-        request: PentestWorkloadRequest,
-        launch_plan: Any,
-        provider_materialization: Any,
-        execution_materialization: Any,
-        provider_lease: Mapping[str, Any],
-        missing_secret_env_keys: Sequence[str],
-    ) -> dict[str, Any]:
-        paths = execution_materialization.runtime_paths
-        provider_snapshot = {
-            "profile_id": provider_materialization.profile_id,
-            "provider_id": provider_materialization.provider_id,
-            "runtime_id": provider_materialization.runtime_id,
-            "force_non_interactive": provider_materialization.force_non_interactive,
-            "env_keys": sorted(provider_materialization.env),
-            "secret_env_keys": sorted(provider_materialization.secret_env),
-            "secret_env_key_count": len(provider_materialization.secret_env),
-            "missing_secret_env_keys": sorted(missing_secret_env_keys),
-            "missing_secret_env_key_count": len(missing_secret_env_keys),
-            "lease": dict(provider_lease),
-            "telemetry_enabled": settings.pentest.telemetry_enabled,
-        }
-        manifest = {
-            "tool_name": "security.pentest.run",
-            "tool_version": "1.0.0",
-            "agent_run_id": request.agent_run_id,
-            "step_id": request.step_id,
-            "attempt": request.attempt,
-            "target": request.target,
-            "scope_artifact_ref": request.scope_artifact_ref,
-            "operation_mode": request.operation_mode,
-            "runner_profile_id": request.runner_profile_id,
-            "execution_profile_ref": request.execution_profile_ref,
-            "time_budget_minutes": request.time_budget_minutes,
-            "evidence_level": request.evidence_level,
-            "launch_plan": {
-                "profile_id": launch_plan.profile_id,
-                "image": launch_plan.image,
-                "network_policy": launch_plan.network_policy,
-                "timeout_seconds": launch_plan.timeout_seconds,
-                "labels": dict(launch_plan.labels),
-            },
-            "runtime_paths": paths.model_dump(mode="json"),
-            "provider_snapshot_ref": f"file:{paths.provider_snapshot_file}",
-        }
-        instruction_file = Path(paths.instruction_file)
-        instruction_file.parent.mkdir(parents=True, exist_ok=True)
-        instruction_file.write_text(
-            redact_sensitive_text(
-                execution_materialization.instruction_bundle.content
-            ),
-            encoding="utf-8",
-        )
-        self._write_pentest_json(paths.input_manifest_file, manifest)
-        if request.approved_scope is not None:
-            self._write_pentest_json(
-                paths.approved_scope_file,
-                request.approved_scope.model_dump(mode="json"),
-            )
-        provider_snapshot_artifact = {
-            "profile_id": str(provider_materialization.profile_id),
-            "provider_id": str(provider_materialization.provider_id),
-            "runtime_id": str(provider_materialization.runtime_id),
-            "force_non_interactive": bool(
-                provider_materialization.force_non_interactive
-            ),
-            "env_keys": sorted(provider_materialization.env),
-            "credential_env_key_count": len(provider_materialization.secret_env),
-            "missing_credential_env_key_count": len(missing_secret_env_keys),
-            "lease": dict(provider_lease),
-            "telemetry_enabled": settings.pentest.telemetry_enabled,
-        }
-        provider_snapshot_file = Path(paths.provider_snapshot_file)
-        provider_snapshot_file.parent.mkdir(parents=True, exist_ok=True)
-        provider_snapshot_file.write_text(
-            json.dumps(provider_snapshot_artifact, sort_keys=True, indent=2) + "\n",
-            encoding="utf-8",
-        )
-        return {
-            "input_manifest_ref": f"file:{paths.input_manifest_file}",
-            "approved_scope_snapshot_ref": f"file:{paths.approved_scope_file}",
-            "provider_snapshot_ref": f"file:{paths.provider_snapshot_file}",
-            "provider_snapshot": provider_snapshot,
-        }
-
-    @staticmethod
-    def _load_pentest_runner_findings(path: str) -> dict[str, Any] | None:
-        try:
-            payload = json.loads(Path(path).read_text(encoding="utf-8"))
-        except (OSError, json.JSONDecodeError):
-            return None
-        if not isinstance(payload, Mapping):
-            return None
-        normalized = redact_sensitive_payload(dict(payload))
-        findings = normalized.get("findings")
-        if not isinstance(findings, list):
-            normalized["findings"] = []
-        else:
-            normalized["findings"] = [
-                item for item in findings if isinstance(item, Mapping)
-            ]
-        summary = normalized.get("summary")
-        if not isinstance(summary, Mapping):
-            confirmed = sum(
-                1
-                for item in normalized["findings"]
-                if str(item.get("confidence") or "") == "confirmed"
-            )
-            high_or_critical = sum(
-                1
-                for item in normalized["findings"]
-                if str(item.get("severity") or "") in {"high", "critical"}
-            )
-            normalized["summary"] = {
-                "findings_count": len(normalized["findings"]),
-                "confirmed_findings_count": confirmed,
-                "high_or_critical_count": high_or_critical,
-            }
-        return dict(normalized)
-
-    async def _publish_pentest_success_artifacts(
-        self,
-        *,
-        request: PentestWorkloadRequest,
-        paths: Any,
-        normalized_findings: Mapping[str, Any],
-        finding_summary: Mapping[str, Any],
-        severity_counts: Mapping[str, int],
-    ) -> dict[str, Any] | None:
-        if (
-            self._artifact_service is None
-            or not hasattr(self._artifact_service, "create")
-            or not hasattr(self._artifact_service, "publish_report_bundle")
-        ):
-            return None
-
-        try:
-            info = temporal_activity.info()
-            namespace = info.namespace
-            workflow_id = info.workflow_id
-            run_id = info.workflow_run_id
-        except Exception:
-            namespace = "default"
-            workflow_id = request.agent_run_id
-            run_id = f"{request.step_id}-{request.attempt}"
-
-        principal = request.principal_id
-        common_metadata = {
-            "producer": "activity:security.pentest.execute",
-            "subject": request.target,
-            "step_id": request.step_id,
-            "attempt": request.attempt,
-            "scope": request.scope_artifact_ref,
-            "sensitivity": "security_restricted",
-        }
-
-        async def publish_file(
-            *,
-            link_type: str,
-            path: str,
-            content_type: str,
-            metadata: Mapping[str, Any],
-            required: bool = True,
-        ) -> dict[str, Any] | None:
-            file_path = Path(path)
-            try:
-                payload = await asyncio.to_thread(file_path.read_bytes)
-            except OSError:
-                if required:
-                    raise
-                return None
-            artifact, _upload = await self._artifact_service.create(
-                principal=principal,
-                content_type=content_type,
-                size_bytes=len(payload),
-                link=ExecutionRef(
-                    namespace=namespace,
-                    workflow_id=workflow_id,
-                    run_id=run_id,
-                    link_type=link_type,
-                    label=str(metadata.get("name") or link_type),
-                ),
-                metadata_json=dict(metadata),
-            )
-            completed = await self._artifact_service.write_complete(
-                artifact_id=artifact.artifact_id,
-                principal=principal,
-                payload=payload,
-                content_type=content_type,
-            )
-            ref = build_artifact_ref(completed)
-            return {"artifact_ref_v": ref.artifact_ref_v, "artifact_id": ref.artifact_id}
-
-        runtime_refs: dict[str, dict[str, Any] | None] = {}
-        runtime_specs = {
-            "runtime.stdout": (
-                paths.stdout_file,
-                "text/plain",
-                "pentest-stdout.log",
-                "runtime_stdout",
-            ),
-            "runtime.stderr": (
-                paths.stderr_file,
-                "text/plain",
-                "pentest-stderr.log",
-                "runtime_stderr",
-            ),
-            "runtime.diagnostics": (
-                paths.diagnostics_file,
-                "application/json",
-                "pentest-diagnostics.json",
-                "runtime_diagnostics",
-            ),
-            "output.provider_snapshot": (
-                paths.provider_snapshot_file,
-                "application/json",
-                "provider-snapshot.json",
-                "provider_snapshot",
-            ),
-        }
-        for link_type, (path, content_type, name, artifact_kind) in runtime_specs.items():
-            runtime_refs[link_type] = await publish_file(
-                link_type=link_type,
-                path=str(path),
-                content_type=content_type,
-                metadata={
-                    **common_metadata,
-                    "name": name,
-                    "artifact_kind": artifact_kind,
-                },
-                required=link_type != "output.provider_snapshot",
-            )
-
-        counts = {
-            "findings_count": int(finding_summary.get("findings_count") or 0),
-            "confirmed_findings_count": int(
-                finding_summary.get("confirmed_findings_count") or 0
-            ),
-            "high_or_critical_count": int(
-                finding_summary.get("high_or_critical_count") or 0
-            ),
-            "severity_counts": dict(severity_counts),
-        }
-        report_metadata = {
-            "artifact_type": "security_pentest_report",
-            "producer": "activity:security.pentest.execute",
-            "subject": request.target,
-            "counts": counts,
-        }
-        workspace_root = Path(paths.workspace_root)
-        report_path = workspace_root / "findings" / "findings.report.md"
-        summary_path = workspace_root / "findings" / "findings.summary.md"
-        evidence_path = workspace_root / "evidence" / "bundle.tar.zst"
-        missing_report_files = [
-            (report_path, "Pentest primary report file was not found"),
-            (summary_path, "Pentest summary report file was not found"),
-            (evidence_path, "Pentest evidence bundle file was not found"),
-        ]
-        for file_path, error_message in missing_report_files:
-            if not file_path.is_file():
-                raise TemporalArtifactValidationError(error_message)
-        primary_payload = await asyncio.to_thread(report_path.read_bytes)
-        summary_payload = await asyncio.to_thread(summary_path.read_bytes)
-        structured_payload = json.dumps(
-            normalized_findings,
-            sort_keys=True,
-        ).encode("utf-8")
-        evidence_payload = await asyncio.to_thread(evidence_path.read_bytes)
-
-        bundle = await self._artifact_service.publish_report_bundle(
-            principal=principal,
-            namespace=namespace,
-            workflow_id=workflow_id,
-            run_id=run_id,
-            report_type="security_pentest_report",
-            report_scope="final",
-            sensitivity="security_restricted",
-            counts=counts,
-            step_id=request.step_id,
-            attempt=request.attempt,
-            scope=request.scope_artifact_ref,
-            primary={
-                "payload": primary_payload,
-                "content_type": "text/markdown",
-                "label": "Pentest primary report",
-                "metadata": {
-                    **report_metadata,
-                    "title": "Pentest report",
-                    "description": "Canonical human-facing Pentest findings report.",
-                    "name": "findings.report.md",
-                    "render_hint": "markdown",
-                },
-            },
-            summary={
-                "payload": summary_payload,
-                "content_type": "text/markdown",
-                "label": "Pentest report summary",
-                "metadata": {
-                    **report_metadata,
-                    "title": "Pentest report summary",
-                    "description": "Short Pentest findings summary.",
-                    "name": "findings.summary.md",
-                    "render_hint": "markdown",
-                },
-            },
-            structured={
-                "payload": structured_payload,
-                "content_type": "application/json",
-                "label": "Pentest structured findings",
-                "metadata": {
-                    **report_metadata,
-                    "title": "Pentest structured findings",
-                    "description": "Machine-readable normalized Pentest findings.",
-                    "name": "findings.normalizer-input.json",
-                    "render_hint": "json",
-                },
-            },
-            evidence=[
-                {
-                    "payload": evidence_payload,
-                    "content_type": "application/zstd",
-                    "label": "Pentest restricted evidence bundle",
-                    "metadata": {
-                        **report_metadata,
-                        "title": "Pentest restricted evidence",
-                        "description": "Restricted Pentest evidence bundle.",
-                        "name": "bundle.tar.zst",
-                        "render_hint": "binary",
-                    },
-                }
-            ],
-        )
-        validate_report_bundle_result(bundle)
-        return {
-            "namespace": namespace,
-            "workflow_id": workflow_id,
-            "run_id": run_id,
-            "runtime_refs": runtime_refs,
-            "report_bundle": bundle,
-            "counts": counts,
-        }
+        self._pentest_activities = TemporalPentestActivities(self)
 
     async def execution_notify_completion(
         self,
@@ -6250,16 +5666,14 @@ class TemporalAgentRuntimeActivities:
     ) -> dict[str, Any]:
         """Registry-dispatched PentestGPT activity boundary (untrusted input).
 
-        This is the entrypoint bound to the ``security.pentest.execute``
-        activity type, so its payload may originate from a caller-supplied
-        plan. The trusted/internal decision is therefore derived from the
-        entrypoint (here, always untrusted) rather than from request inputs:
-        an inline ``approved_scope`` is rejected and the scope is always loaded
-        from the artifact store. Trusted internal callers must use
-        :meth:`_security_pentest_execute_trusted_internal` instead.
+        Thin delegate to :class:`TemporalPentestActivities`. This is the
+        entrypoint bound to the ``security.pentest.execute`` activity type, so
+        its payload may originate from a caller-supplied plan and is always
+        treated as untrusted: an inline ``approved_scope`` is rejected and the
+        scope is always loaded from the artifact store.
         """
 
-        return await self._run_security_pentest(payload, trusted_internal=False)
+        return await self._pentest_activities.security_pentest_execute(payload)
 
     async def _security_pentest_execute_trusted_internal(
         self,
@@ -6268,936 +5682,15 @@ class TemporalAgentRuntimeActivities:
     ) -> dict[str, Any]:
         """Internal entrypoint that may honor an inline ``approved_scope``.
 
-        This method is intentionally **not** registered in
-        ``_ACTIVITY_HANDLER_ATTRS``, so it cannot be reached through
-        registry/plan dispatch. The trusted decision comes from the fact that
-        only trusted workflow-internal code can call this entrypoint, never
-        from a ``trusted_internal_execution`` flag in the request payload.
+        Thin delegate to :class:`TemporalPentestActivities`. Intentionally
+        **not** registered in ``_ACTIVITY_HANDLER_ATTRS`` so it cannot be
+        reached through registry/plan dispatch; only trusted workflow-internal
+        code can call it.
         """
 
-        return await self._run_security_pentest(payload, trusted_internal=True)
-
-    async def _run_security_pentest(
-        self,
-        payload: Mapping[str, Any],
-        /,
-        *,
-        trusted_internal: bool,
-    ) -> dict[str, Any]:
-        """Shared policy-gated PentestGPT workload attempt.
-
-        The full PentestGPT runner is intentionally implemented separately from
-        the registry contract. Keeping this activity registered lets workflow
-        validation and worker startup recognize the curated route without
-        silently falling back to the generic tool executor.
-        """
-
-        if self._workflow_docker_mode == "disabled":
-            raise _docker_workflows_disabled_failure()
-        raw_payload = dict(payload)
-        nested_request = raw_payload.get("request")
-        if nested_request is not None:
-            request_payload = dict(nested_request)
-            publication_source = request_payload
-        else:
-            request_payload = raw_payload
-            publication_source = raw_payload
-        publication_keys = {
-            "findings",
-            "provider_snapshot_available",
-            "wrapper_log_available",
-            "summary_text",
-            "stdout_preview",
-            "stderr_preview",
-            "progress_annotations",
-            "publication_errors",
-        }
-        policy_keys = {"pentest_enabled", "execution_policy"}
-        publication_payload = {
-            key: publication_source[key]
-            for key in publication_keys
-            if key in publication_source
-        }
-        policy_payload = {
-            key: publication_source[key]
-            for key in policy_keys
-            if key in publication_source
-        }
-        request_payload = {
-            key: value
-            for key, value in request_payload.items()
-            if key not in publication_keys and key not in policy_keys
-        }
-        execution_policy = PentestExecutionPolicy()
-
-        def progress(phase: str, message: str) -> dict[str, Any]:
-            return build_pentest_progress_annotation(
-                phase=phase,
-                message=message,
-            ).model_dump(mode="json")
-
-        def cleanup(
-            *,
-            terminal_reason: str,
-            last_phase: str,
-            launched: bool,
-            provider_lease_released: bool = False,
-            original_failure_summary: str | None = None,
-            stdout_ref: str | None = None,
-            stderr_ref: str | None = None,
-            diagnostics_ref: str | None = None,
-            evidence_refs: Sequence[str] = (),
-        ) -> dict[str, Any]:
-            return build_pentest_terminal_cleanup_result(
-                terminal_reason=terminal_reason,  # type: ignore[arg-type]
-                last_phase=last_phase,
-                graceful_termination_attempted=launched,
-                kill_escalated=False,
-                stdout_artifact_ref=stdout_ref,
-                stderr_artifact_ref=stderr_ref,
-                diagnostics_artifact_ref=diagnostics_ref,
-                partial_evidence_artifact_refs=tuple(evidence_refs),
-                provider_lease_released=provider_lease_released,
-                container_removed=launched,
-                original_failure_summary=original_failure_summary,
-            ).model_dump(mode="json")
-
-        def structured_failure(
-            *,
-            status: str,
-            target: str | None,
-            failure_kind: str,
-            interaction_state: str,
-            phase: str,
-            reason: str,
-            diagnostics: Mapping[str, Any] | None = None,
-            launched: bool = False,
-            provider_lease_released: bool = False,
-            stdout_ref: str | None = None,
-            stderr_ref: str | None = None,
-            diagnostics_ref: str | None = None,
-            evidence_refs: Sequence[str] = (),
-        ) -> dict[str, Any]:
-            redacted_reason = redact_pentest_human_text(reason)
-            return {
-                "status": status,
-                "target": target,
-                "execution_policy": execution_policy.model_dump(mode="json"),
-                "failure_classification": classify_pentest_failure(
-                    failure_kind=failure_kind,  # type: ignore[arg-type]
-                    interaction_state=interaction_state,  # type: ignore[arg-type]
-                    phase=phase,
-                    diagnostics={
-                        "reason": redacted_reason,
-                        **dict(diagnostics or {}),
-                    },
-                ).model_dump(mode="json"),
-                "progress_annotation": progress(phase, redacted_reason),
-                "terminal_cleanup": cleanup(
-                    terminal_reason="failure",
-                    last_phase=phase,
-                    launched=launched,
-                    provider_lease_released=provider_lease_released,
-                    original_failure_summary=redacted_reason,
-                    stdout_ref=stdout_ref,
-                    stderr_ref=stderr_ref,
-                    diagnostics_ref=diagnostics_ref,
-                    evidence_refs=evidence_refs,
-                ),
-                "diagnostics": {
-                    "reason": redacted_reason,
-                    **dict(diagnostics or {}),
-                },
-            }
-
-        lease_runtime_id: str | None = None
-        lease_profile_id: str | None = None
-        lease_owner: str | None = None
-        lease_id: str | None = None
-        lease_release_attempted = False
-
-        async def release_provider_lease() -> bool:
-            nonlocal lease_release_attempted
-            if not all(
-                (lease_id, lease_runtime_id, lease_profile_id, lease_owner)
-            ):
-                return False
-            if lease_release_attempted:
-                return True
-            lease_release_attempted = True
-            await self._pentest_provider_lease_manager.release(
-                runtime_id=lease_runtime_id,
-                profile_id=lease_profile_id,
-                owner=lease_owner,
-                lease_id=lease_id,
-            )
-            return True
-
-        enabled = _coerce_bool(
-            policy_payload.get(
-                "pentest_enabled",
-                settings.pentest.enabled,
-            ),
-            default=False,
+        return await self._pentest_activities._security_pentest_execute_trusted_internal(
+            payload
         )
-        try:
-            pre_validation_attempt = int(request_payload.get("attempt") or 1)
-        except (TypeError, ValueError):
-            pre_validation_attempt = 1
-        emit_pentest_activity_heartbeat(
-            phase="validating_scope",
-            agent_run_id=str(request_payload.get("agent_run_id") or "") or None,
-            step_id=str(request_payload.get("step_id") or "") or None,
-            attempt=pre_validation_attempt,
-            message="Validating Pentest scope and policy.",
-        )
-        if not enabled:
-            emit_pentest_activity_heartbeat(
-                phase="cleanup",
-                agent_run_id=str(request_payload.get("agent_run_id") or "") or None,
-                step_id=str(request_payload.get("step_id") or "") or None,
-                attempt=pre_validation_attempt,
-                message="Pentest execution denied before launch.",
-                metadata={"terminal_reason": "failure"},
-            )
-            return structured_failure(
-                status="policy_denied",
-                target=str(request_payload.get("target") or "") or None,
-                failure_kind="policy_denied",
-                interaction_state="pre_interaction",
-                phase="validating_scope",
-                reason="Pentest execution is disabled by operator policy.",
-                diagnostics={"policy": "MOONMIND_PENTEST_ENABLED=false"},
-            )
-
-        try:
-            if trusted_internal:
-                request_payload["trusted_internal_execution"] = True
-                if not request_payload.get("scope_artifact_ref"):
-                    request_payload["scope_artifact_ref"] = (
-                        "trusted-internal-inline-scope"
-                    )
-            else:
-                # Untrusted callers can never assert trusted-internal execution
-                # or supply their own inline scope, regardless of what the
-                # request payload claims. Drop any caller-supplied trust flag
-                # and require an artifact-backed approved scope.
-                request_payload.pop("trusted_internal_execution", None)
-                if request_payload.get("approved_scope") is not None:
-                    raise PentestScopeValidationError(
-                        error_code="INVALID_SCOPE",
-                        reasons=("inline_scope_requires_trusted_internal_execution",),
-                        diagnostics={
-                            "scope_artifact_ref": request_payload.get(
-                                "scope_artifact_ref"
-                            ),
-                        },
-                    )
-                approved_scope = await self._load_pentest_approved_scope(
-                    request_payload
-                )
-                request_payload["approved_scope"] = approved_scope.model_dump(
-                    mode="json"
-                )
-                request_payload["scope_artifact_loaded"] = True
-            request = PentestWorkloadRequest.model_validate(request_payload)
-            if request.approved_scope is None:
-                raise PentestScopeValidationError(
-                    error_code="INVALID_SCOPE",
-                    reasons=("missing_approved_scope",),
-                    diagnostics={"scope_artifact_ref": request.scope_artifact_ref},
-                )
-            self._validate_pentest_request_policy(request, request.approved_scope)
-            self._validate_pentest_runner_profiles_registered()
-            launch_plan = build_pentest_launch_plan(request)
-            provider_profile = resolve_pentest_provider_profile(
-                execution_profile_ref=request.execution_profile_ref,
-                provider_selector=request.provider_selector,
-                runtime_state=request.provider_runtime_state,
-            )
-            lease_runtime_id = provider_profile.runtime_id
-            lease_profile_id = provider_profile.profile_id
-            lease_owner = _pentest_provider_lease_owner(
-                agent_run_id=request.agent_run_id,
-                step_id=request.step_id,
-                attempt=request.attempt,
-            )
-            emit_pentest_activity_heartbeat(
-                phase="waiting_for_profile_slot",
-                agent_run_id=request.agent_run_id,
-                step_id=request.step_id,
-                attempt=request.attempt,
-                message="Waiting for Pentest provider profile capacity.",
-                metadata={
-                    "runtime_id": lease_runtime_id,
-                    "profile_id": lease_profile_id,
-                },
-            )
-            try:
-                lease_id = await self._pentest_provider_lease_manager.acquire(
-                    runtime_id=lease_runtime_id,
-                    profile_id=lease_profile_id,
-                    owner=lease_owner,
-                    metadata=_pentest_provider_lease_safe_metadata(
-                        request,
-                        runtime_id=lease_runtime_id,
-                        profile_id=lease_profile_id,
-                    ),
-                )
-            except Exception as exc:
-                return structured_failure(
-                    status="provider_capacity_failed",
-                    target=request.target,
-                    failure_kind="provider_capacity",
-                    interaction_state="pre_interaction",
-                    phase="waiting_for_profile_slot",
-                    reason=str(exc),
-                    diagnostics={"provider_capacity": "lease_manager_failure"},
-                )
-            provider_materialization = materialize_pentest_provider_profile(
-                provider_profile
-            )
-            execution_materialization = build_pentest_execution_materialization(
-                request
-            )
-            resolved_secret_env, missing_secret_env_keys = (
-                await self._resolve_pentest_secret_env(provider_materialization)
-            )
-        except (
-            PentestScopeValidationError,
-            PentestLaunchPolicyError,
-            PentestProviderMaterializationError,
-            ValidationError,
-        ) as exc:
-            provider_lease_released = False
-            if lease_id:
-                try:
-                    provider_lease_released = await release_provider_lease()
-                except Exception as release_exc:
-                    logger.warning(
-                        "failed to release Pentest provider lease after validation failure: %s",
-                        release_exc,
-                    )
-            diagnostics = dict(getattr(exc, "diagnostics", {}) or {})
-            reason_code = getattr(exc, "reason", None)
-            if reason_code:
-                diagnostics["policy_reason"] = str(reason_code)
-            return structured_failure(
-                status="validation_failed",
-                target=str(request_payload.get("target") or "") or None,
-                failure_kind="policy_denied",
-                interaction_state="pre_interaction",
-                phase="validating_scope",
-                reason=str(exc),
-                diagnostics=diagnostics if isinstance(diagnostics, Mapping) else {},
-                provider_lease_released=provider_lease_released,
-            )
-        output = launch_plan.to_activity_output(request)
-        output["provider_profile"] = provider_materialization.model_dump(mode="json")
-        provider_lease = pentest_provider_lease_metadata(provider_profile)
-        provider_lease["lease_id"] = lease_id
-        provider_lease["lease_owner"] = lease_owner
-        provider_lease["metadata"] = _pentest_provider_lease_safe_metadata(
-            request,
-            runtime_id=provider_profile.runtime_id,
-            profile_id=provider_profile.profile_id,
-        )
-        output["provider_lease"] = provider_lease
-        execution_metadata = execution_materialization.model_dump(mode="json")
-        execution_metadata["instruction_bundle"].pop("content", None)
-        execution_metadata["instruction_bundle"].pop("objective", None)
-        output.update(execution_metadata)
-        emit_pentest_activity_heartbeat(
-            phase="materializing_inputs",
-            agent_run_id=request.agent_run_id,
-            step_id=request.step_id,
-            attempt=request.attempt,
-            message="Materializing Pentest input files.",
-        )
-        try:
-            output.update(
-                self._materialize_pentest_input_files(
-                    request=request,
-                    launch_plan=launch_plan,
-                    provider_materialization=provider_materialization,
-                    execution_materialization=execution_materialization,
-                    provider_lease=provider_lease,
-                    missing_secret_env_keys=missing_secret_env_keys,
-                )
-            )
-        except OSError as exc:
-            provider_lease_released = False
-            try:
-                provider_lease_released = await release_provider_lease()
-            except Exception as release_exc:
-                logger.warning(
-                    "failed to release Pentest provider lease after materialization failure: %s",
-                    release_exc,
-                )
-            output.update(
-                structured_failure(
-                    status="failed",
-                    target=request.target,
-                    failure_kind="transient_infrastructure",
-                    interaction_state="pre_interaction",
-                    phase="materializing_inputs",
-                    reason=str(exc),
-                    diagnostics={"materialization": "input_file_write_failed"},
-                    provider_lease_released=provider_lease_released,
-                )
-            )
-            return output
-        publication = build_pentest_publication_result(
-            request,
-            findings=list(publication_payload.get("findings") or ()),
-            provider_snapshot_available=True,
-            wrapper_log_available=_coerce_bool(
-                publication_payload.get("wrapper_log_available"),
-                default=False,
-            ),
-            summary_text=publication_payload.get("summary_text"),
-            stdout_preview=publication_payload.get("stdout_preview"),
-            stderr_preview=publication_payload.get("stderr_preview"),
-            progress_annotations=list(
-                publication_payload.get("progress_annotations") or ()
-            ),
-            errors=list(publication_payload.get("publication_errors") or ()),
-        ).model_dump(mode="json")
-        output.update(publication)
-        progress_annotations = [
-            progress(phase, f"Pentest phase {phase}.")
-            for phase in PENTEST_HEARTBEAT_PHASES
-        ]
-        output["progress_annotation"] = progress_annotations[-1]
-        output["progress"] = progress_annotations
-        provider_failure = request.provider_failure or {}
-        failure_category = str(provider_failure.get("category") or "").strip()
-        if failure_category in {"provider_429", "provider_quota"}:
-            cooldown = build_pentest_provider_cooldown_diagnostic(
-                profile_id=provider_profile.profile_id,
-                cooldown_seconds=provider_profile.cooldown_after_429_seconds,
-                reason=failure_category,
-            ).model_dump(mode="json")
-            try:
-                await self._pentest_provider_lease_manager.record_cooldown(
-                    runtime_id=provider_profile.runtime_id,
-                    profile_id=provider_profile.profile_id,
-                    owner=lease_owner or "",
-                    cooldown_seconds=provider_profile.cooldown_after_429_seconds,
-                    reason=failure_category,
-                )
-            except Exception as cooldown_exc:
-                logger.warning(
-                    "failed to record Pentest provider cooldown: %s",
-                    cooldown_exc,
-                )
-            provider_lease_released = False
-            try:
-                provider_lease_released = await release_provider_lease()
-            except Exception as release_exc:
-                logger.warning(
-                    "failed to release Pentest provider lease after cooldown: %s",
-                    release_exc,
-                )
-            output.update(
-                structured_failure(
-                    status="provider_cooldown",
-                    target=request.target,
-                    failure_kind="provider_error",
-                    interaction_state="pre_interaction",
-                    phase="waiting_for_profile_slot",
-                    reason=failure_category,
-                    diagnostics={"provider_cooldown": cooldown},
-                    provider_lease_released=provider_lease_released,
-                )
-            )
-            output["provider_cooldown"] = cooldown
-            output["provider_lease"] = pentest_provider_lease_metadata(
-                provider_profile, release_required=True
-            )
-            output["provider_lease"]["lease_id"] = lease_id
-            output["provider_lease"]["lease_owner"] = lease_owner
-            output["provider_lease"]["metadata"] = _pentest_provider_lease_safe_metadata(
-                request,
-                runtime_id=provider_profile.runtime_id,
-                profile_id=provider_profile.profile_id,
-            )
-            return output
-
-        workload_payload = {
-            "profileId": launch_plan.profile_id,
-            "agentRunId": request.agent_run_id,
-            "stepId": request.step_id,
-            "attempt": request.attempt,
-            "toolName": "security.pentest.run",
-            "runtimeId": PENTEST_RUNTIME_ID,
-            "repoDir": request.repo_dir or launch_plan.workdir,
-            "artifactsDir": request.artifacts_dir,
-            "command": list(execution_materialization.wrapper_invocation.command),
-            "envOverrides": {
-                **provider_materialization.env,
-                **resolved_secret_env,
-                **execution_materialization.wrapper_invocation.env,
-            },
-            "timeoutSeconds": launch_plan.timeout_seconds,
-            "resources": launch_plan.resources,
-            "declaredOutputs": {
-                "report.primary": "pentest/findings/findings.report.md",
-                "report.summary": "pentest/findings/findings.summary.md",
-                "report.structured": "pentest/findings/findings.normalizer-input.json",
-                "report.evidence": "pentest/evidence/bundle.tar.zst",
-            },
-        }
-        workload_request = parse_workload_request(workload_payload)
-        try:
-            if self._workload_launcher is None:
-                emit_pentest_activity_heartbeat(
-                    phase="launching_container",
-                    agent_run_id=request.agent_run_id,
-                    step_id=request.step_id,
-                    attempt=request.attempt,
-                    message="Using built-in Pentest workload fallback.",
-                )
-                emit_pentest_activity_heartbeat(
-                    phase="running",
-                    agent_run_id=request.agent_run_id,
-                    step_id=request.step_id,
-                    attempt=request.attempt,
-                    message="Pentest workload fallback completed.",
-                )
-                workload_result = WorkloadResult.model_validate(
-                    {
-                        "requestId": (
-                            f"pentest-{request.agent_run_id}-{request.step_id}-"
-                            f"{request.attempt}"
-                        ),
-                        "profileId": launch_plan.profile_id,
-                        "status": "succeeded",
-                        "exitCode": 0,
-                        "stdoutRef": f"file:{execution_materialization.runtime_paths.stdout_file}",
-                        "stderrRef": f"file:{execution_materialization.runtime_paths.stderr_file}",
-                        "diagnosticsRef": f"file:{execution_materialization.runtime_paths.diagnostics_file}",
-                        "outputRefs": {
-                            "report.primary": _artifact_ref_for_pentest_name(
-                                publication, "report.primary"
-                            ),
-                            "report.summary": _artifact_ref_for_pentest_name(
-                                publication, "report.summary"
-                            ),
-                            "report.structured": _artifact_ref_for_pentest_name(
-                                publication, "report.structured"
-                            ),
-                            "report.evidence": _artifact_ref_for_pentest_name(
-                                publication, "report.evidence"
-                            ),
-                        },
-                    }
-                )
-            else:
-                if self._workload_registry is None:
-                    raise TemporalActivityRuntimeError(
-                        "workload registry is required to launch pentest workloads"
-                    )
-                validated_workload_request = self._workload_registry.validate_request(
-                    workload_request,
-                    workflow_docker_mode=self._workflow_docker_mode,
-                )
-                emit_pentest_activity_heartbeat(
-                    phase="launching_container",
-                    agent_run_id=request.agent_run_id,
-                    step_id=request.step_id,
-                    attempt=request.attempt,
-                    message="Launching Pentest workload container.",
-                    metadata={
-                        "profile_id": launch_plan.profile_id,
-                        "container_name": launch_plan.container_name,
-                    },
-                )
-                workload_result = (
-                    await _supervise_pentest_workload_with_activity_heartbeats(
-                        self._workload_launcher,
-                        validated_workload_request,
-                        request=request,
-                        timeout_seconds=float(launch_plan.timeout_seconds),
-                        heartbeat_interval_seconds=(
-                            _PENTEST_RUNNING_HEARTBEAT_INTERVAL_SECONDS
-                        ),
-                    )
-                )
-        except asyncio.CancelledError:
-            provider_lease_released = False
-            try:
-                provider_lease_released = await release_provider_lease()
-            except Exception as release_exc:
-                logger.warning(
-                    "failed to release Pentest provider lease after workload cancellation: %s",
-                    release_exc,
-                )
-            emit_pentest_activity_heartbeat(
-                phase="cleanup",
-                agent_run_id=request.agent_run_id,
-                step_id=request.step_id,
-                attempt=request.attempt,
-                message="Pentest workload cancellation cleanup completed.",
-                metadata={
-                    "terminal_reason": "cancellation",
-                    "provider_lease_released": provider_lease_released,
-                    "container_removed": True,
-                },
-            )
-            output.update(
-                structured_failure(
-                    status="canceled",
-                    target=request.target,
-                    failure_kind="runtime_action",
-                    interaction_state="post_interaction",
-                    phase="running",
-                    reason="Pentest workload cancelled.",
-                    launched=True,
-                    provider_lease_released=provider_lease_released,
-                )
-            )
-            output["terminal_cleanup"]["terminal_reason"] = "cancellation"
-            output["terminal_cleanup"]["graceful_termination_attempted"] = True
-            output["terminal_cleanup"]["container_removed"] = True
-            return output
-        except Exception as exc:
-            provider_lease_released = False
-            try:
-                provider_lease_released = await release_provider_lease()
-            except Exception as release_exc:
-                logger.warning(
-                    "failed to release Pentest provider lease after workload failure: %s",
-                    release_exc,
-                )
-            output.update(
-                structured_failure(
-                    status="failed",
-                    target=request.target,
-                    failure_kind="runtime_action",
-                    interaction_state="post_interaction",
-                    phase="running",
-                    reason=str(exc),
-                    launched=True,
-                    provider_lease_released=provider_lease_released,
-                )
-            )
-            emit_pentest_activity_heartbeat(
-                phase="cleanup",
-                agent_run_id=request.agent_run_id,
-                step_id=request.step_id,
-                attempt=request.attempt,
-                message="Pentest workload failure cleanup completed.",
-                metadata={
-                    "terminal_reason": "failure",
-                    "provider_lease_released": provider_lease_released,
-                    "container_removed": True,
-                },
-            )
-            return output
-
-        workload_dump = workload_result.model_dump(mode="json", by_alias=True)
-        output["workload_result"] = workload_dump
-        workload_cleanup = (
-            workload_result.metadata.get("cleanup")
-            if isinstance(workload_result.metadata, Mapping)
-            else None
-        )
-        if not isinstance(workload_cleanup, Mapping):
-            workload_cleanup = {}
-        runner_findings = self._load_pentest_runner_findings(
-            execution_materialization.runtime_paths.normalizer_input_file
-        )
-        if runner_findings is not None:
-            publication["normalized_findings"] = runner_findings
-            output["normalized_findings"] = runner_findings
-        emit_pentest_activity_heartbeat(
-            phase="publishing_artifacts",
-            agent_run_id=request.agent_run_id,
-            step_id=request.step_id,
-            attempt=request.attempt,
-            message="Publishing Pentest artifact metadata.",
-            metadata={
-                "artifact_count": len(
-                    publication.get("artifact_publication", {}).get("artifacts", ())
-                ),
-                "status": publication.get("artifact_publication", {}).get("status"),
-            },
-        )
-        emit_pentest_activity_heartbeat(
-            phase="normalizing_findings",
-            agent_run_id=request.agent_run_id,
-            step_id=request.step_id,
-            attempt=request.attempt,
-            message="Normalizing Pentest findings.",
-            metadata={
-                "findings_count": publication["normalized_findings"]["summary"][
-                    "findings_count"
-                ],
-            },
-        )
-        output_refs = workload_result.output_refs or {}
-        stdout_ref = workload_result.stdout_ref or _artifact_ref_for_pentest_name(
-            publication, "runtime.stdout"
-        )
-        stderr_ref = workload_result.stderr_ref or _artifact_ref_for_pentest_name(
-            publication, "runtime.stderr"
-        )
-        diagnostics_ref = (
-            workload_result.diagnostics_ref
-            or _artifact_ref_for_pentest_name(publication, "runtime.diagnostics")
-        )
-        summary_ref = output_refs.get("report.summary") or _artifact_ref_for_pentest_name(
-            publication, "report.summary"
-        )
-        primary_ref = output_refs.get("report.primary") or _artifact_ref_for_pentest_name(
-            publication, "report.primary"
-        )
-        structured_ref = output_refs.get(
-            "report.structured"
-        ) or _artifact_ref_for_pentest_name(publication, "report.structured")
-        evidence_ref = output_refs.get("report.evidence") or _artifact_ref_for_pentest_name(
-            publication, "report.evidence"
-        )
-        finding_summary = publication["normalized_findings"]["summary"]
-        severity_counts: dict[str, int] = {}
-        for finding in publication["normalized_findings"].get("findings", []):
-            if not isinstance(finding, Mapping):
-                continue
-            severity = str(finding.get("severity") or "informational").strip()
-            severity_counts[severity] = severity_counts.get(severity, 0) + 1
-
-        if workload_result.status != "succeeded":
-            provider_lease_released = False
-            try:
-                provider_lease_released = await release_provider_lease()
-            except Exception as release_exc:
-                logger.warning(
-                    "failed to release Pentest provider lease after terminal workload status: %s",
-                    release_exc,
-                )
-            terminal_reason = (
-                "timeout"
-                if workload_result.status == "timed_out"
-                else "cancellation"
-                if workload_result.status == "canceled"
-                else "failure"
-            )
-            output.update(
-                structured_failure(
-                    status=workload_result.status.replace("_", "-")
-                    if workload_result.status == "timed_out"
-                    else workload_result.status,
-                    target=request.target,
-                    failure_kind="runtime_action",
-                    interaction_state="post_interaction",
-                    phase="running",
-                    reason=workload_result.timeout_reason
-                    or f"Pentest workload ended with status {workload_result.status}.",
-                    launched=True,
-                    provider_lease_released=provider_lease_released,
-                    stdout_ref=stdout_ref,
-                    stderr_ref=stderr_ref,
-                    diagnostics_ref=diagnostics_ref,
-                    evidence_refs=(evidence_ref,),
-                )
-            )
-            output["terminal_cleanup"]["terminal_reason"] = terminal_reason
-            output["terminal_cleanup"]["graceful_termination_attempted"] = True
-            output["terminal_cleanup"]["kill_escalated"] = bool(
-                workload_cleanup.get(
-                    "kill_escalated",
-                    workload_cleanup.get("killEscalated", False),
-                )
-            )
-            output["terminal_cleanup"]["container_removed"] = bool(
-                workload_cleanup.get(
-                    "container_removed",
-                    workload_cleanup.get("containerRemoved", True),
-                )
-            )
-            cleanup_errors = workload_cleanup.get(
-                "cleanup_errors",
-                workload_cleanup.get("cleanupErrors", ()),
-            )
-            if isinstance(cleanup_errors, Sequence) and not isinstance(
-                cleanup_errors, (str, bytes)
-            ):
-                output["terminal_cleanup"]["cleanup_errors"] = [
-                    redact_pentest_human_text(str(item))
-                    for item in cleanup_errors
-                ]
-            emit_pentest_activity_heartbeat(
-                phase="cleanup",
-                agent_run_id=request.agent_run_id,
-                step_id=request.step_id,
-                attempt=request.attempt,
-                message="Pentest terminal workload cleanup completed.",
-                metadata={
-                    "terminal_reason": terminal_reason,
-                    "provider_lease_released": provider_lease_released,
-                    "container_removed": output["terminal_cleanup"][
-                        "container_removed"
-                    ],
-                    "kill_escalated": output["terminal_cleanup"]["kill_escalated"],
-                },
-            )
-            return output
-
-        provider_lease_released = False
-        try:
-            provider_lease_released = await release_provider_lease()
-        except Exception as release_exc:
-            logger.warning(
-                "failed to release Pentest provider lease after success: %s",
-                release_exc,
-            )
-        published_artifacts = await self._publish_pentest_success_artifacts(
-            request=request,
-            paths=execution_materialization.runtime_paths,
-            normalized_findings=publication["normalized_findings"],
-            finding_summary=finding_summary,
-            severity_counts=severity_counts,
-        )
-        report_bundle = None
-        counts = {
-            "findings_count": finding_summary.get("findings_count", 0),
-            "confirmed_findings_count": finding_summary.get(
-                "confirmed_findings_count", 0
-            ),
-            "high_or_critical_count": finding_summary.get(
-                "high_or_critical_count", 0
-            ),
-            "severity_counts": severity_counts,
-        }
-        if published_artifacts is not None:
-            runtime_refs = published_artifacts["runtime_refs"]
-            stdout_artifact = runtime_refs.get("runtime.stdout")
-            stderr_artifact = runtime_refs.get("runtime.stderr")
-            diagnostics_artifact = runtime_refs.get("runtime.diagnostics")
-            provider_snapshot_artifact = runtime_refs.get("output.provider_snapshot")
-            report_bundle = published_artifacts["report_bundle"]
-            counts = dict(published_artifacts["counts"])
-
-            if isinstance(stdout_artifact, Mapping):
-                stdout_ref = str(stdout_artifact.get("artifact_id") or stdout_ref)
-            if isinstance(stderr_artifact, Mapping):
-                stderr_ref = str(stderr_artifact.get("artifact_id") or stderr_ref)
-            if isinstance(diagnostics_artifact, Mapping):
-                diagnostics_ref = str(
-                    diagnostics_artifact.get("artifact_id") or diagnostics_ref
-                )
-            primary_report_ref = report_bundle.get("primary_report_ref")
-            summary_report_ref = report_bundle.get("summary_ref")
-            structured_report_ref = report_bundle.get("structured_ref")
-            evidence_report_refs = report_bundle.get("evidence_refs") or []
-            if isinstance(primary_report_ref, Mapping):
-                primary_ref = str(primary_report_ref.get("artifact_id") or primary_ref)
-            if isinstance(summary_report_ref, Mapping):
-                summary_ref = str(summary_report_ref.get("artifact_id") or summary_ref)
-            if isinstance(structured_report_ref, Mapping):
-                structured_ref = str(
-                    structured_report_ref.get("artifact_id") or structured_ref
-                )
-            if evidence_report_refs and isinstance(evidence_report_refs[0], Mapping):
-                evidence_ref = str(
-                    evidence_report_refs[0].get("artifact_id") or evidence_ref
-                )
-            if isinstance(provider_snapshot_artifact, Mapping):
-                provider_snapshot_ref = str(
-                    provider_snapshot_artifact.get("artifact_id") or ""
-                )
-            else:
-                provider_snapshot_ref = None
-        else:
-            provider_snapshot_ref = _artifact_ref_for_pentest_name(
-                publication, "output.provider_snapshot"
-            )
-            report_bundle = {
-                "report_bundle_v": 1,
-                "primary_report_ref": {
-                    "artifact_ref_v": 1,
-                    "artifact_id": primary_ref,
-                },
-                "summary_ref": {"artifact_ref_v": 1, "artifact_id": summary_ref},
-                "structured_ref": {
-                    "artifact_ref_v": 1,
-                    "artifact_id": structured_ref,
-                },
-                "evidence_refs": (
-                    [{"artifact_ref_v": 1, "artifact_id": evidence_ref}]
-                    if evidence_ref
-                    else []
-                ),
-                "report_type": "security_pentest_report",
-                "report_scope": "final",
-                "sensitivity": "security_restricted",
-                "counts": counts,
-            }
-        result = PentestWorkloadResult(
-            status="completed",
-            target=request.target,
-            runner_profile_id=request.runner_profile_id,
-            execution_profile_ref=request.execution_profile_ref,
-            stdout_artifact_ref=stdout_ref,
-            stderr_artifact_ref=stderr_ref,
-            diagnostics_artifact_ref=diagnostics_ref,
-            summary_artifact_ref=summary_ref,
-            findings_artifact_ref=structured_ref,
-            evidence_bundle_artifact_ref=evidence_ref,
-            provider_snapshot_artifact_ref=provider_snapshot_ref,
-            findings_count=finding_summary.get("findings_count", 0),
-            confirmed_findings_count=finding_summary.get(
-                "confirmed_findings_count", 0
-            ),
-        )
-        output.update(result.model_dump(mode="json"))
-        output["status"] = "completed"
-        output["report_bundle_v"] = 1
-        output["primary_report_ref"] = primary_ref
-        output["summary_ref"] = summary_ref
-        output["structured_ref"] = structured_ref
-        output["evidence_refs"] = [evidence_ref] if evidence_ref else []
-        output["report_type"] = "security_pentest_report"
-        output["report_scope"] = "final"
-        output["sensitivity"] = {
-            "classification": "security_restricted",
-            "redaction_level": "restricted",
-            "contains_pentest_evidence": True,
-            "raw_observability_artifacts": [
-                "runtime.stdout",
-                "runtime.stderr",
-                "runtime.diagnostics",
-            ],
-        }
-        output["severity_counts"] = severity_counts
-        output["counts"] = counts
-        output["report_bundle"] = report_bundle
-        validate_report_bundle_result(output["report_bundle"])
-        output["execution_policy"] = execution_policy.model_dump(mode="json")
-        output["terminal_cleanup"] = cleanup(
-            terminal_reason="success",
-            last_phase="cleanup",
-            launched=True,
-            provider_lease_released=provider_lease_released,
-            stdout_ref=stdout_ref,
-            stderr_ref=stderr_ref,
-            diagnostics_ref=diagnostics_ref,
-            evidence_refs=(evidence_ref,),
-        )
-        emit_pentest_activity_heartbeat(
-            phase="cleanup",
-            agent_run_id=request.agent_run_id,
-            step_id=request.step_id,
-            attempt=request.attempt,
-            message="Pentest workload cleanup completed.",
-            metadata={
-                "terminal_reason": "success",
-                "provider_lease_released": provider_lease_released,
-                "container_removed": output["terminal_cleanup"]["container_removed"],
-            },
-        )
-        return output
 
     async def agent_runtime_publish_artifacts(
         self,
