+++ tests/integration/temporal/test_pentest_docker_e2e.py
@@ -1,0 +1,177 @@
+"""Phase 2/3: real Docker workload launcher + provider-manager E2E.
+
+These tests exercise the curated PentestGPT activity through the **real**
+``DockerWorkloadLauncher`` and the real ``RunnerProfileRegistry`` loaded from
+``config/workloads/default-runner-profiles.yaml`` — not the fake/file-writing
+launchers used by the hermetic contract tests.
+
+They require a Docker daemon and the curated runner image, so they are marked
+``integration`` only (NOT ``integration_ci``) per the repo test taxonomy:
+local/staging dev only, excluded from required CI. They additionally self-skip
+unless explicitly opted in, so they never run by accident:
+
+* ``MOONMIND_PENTEST_E2E=1`` and a Docker daemon  -> Phase 2 docker E2E
+* ``MOONMIND_PENTEST_PROVIDER_E2E=1``             -> Phase 3 provider-manager E2E
+
+Run locally with, e.g.::
+
+    MOONMIND_PENTEST_E2E=1 \
+    MOONMIND_PENTEST_RUNNER_IMAGE=ghcr.io/moonladderstudios/moonmind-pentestgpt:1.0 \
+    python -m pytest tests/integration/temporal/test_pentest_docker_e2e.py -m integration
+"""
+
+from __future__ import annotations
+
+import os
+import shutil
+from pathlib import Path
+
+import pytest
+
+from moonmind.config import settings
+from moonmind.workloads.registry import RunnerProfileRegistry
+
+# Reuse the known-good helpers/fixtures from the unit activity-runtime suite.
+from tests.unit.workflows.temporal.test_activity_runtime import (  # noqa: E501
+    LocalTemporalArtifactStore,
+    TemporalAgentRuntimeActivities,
+    TemporalArtifactRepository,
+    TemporalArtifactService,
+    _approved_pentest_scope,
+    _pentest_activity_payload,
+    temporal_db,
+)
+
+pytestmark = [pytest.mark.asyncio, pytest.mark.integration]
+
+_DOCKER_AVAILABLE = shutil.which("docker") is not None
+_E2E_ENABLED = os.environ.get("MOONMIND_PENTEST_E2E") == "1"
+_PROVIDER_E2E_ENABLED = os.environ.get("MOONMIND_PENTEST_PROVIDER_E2E") == "1"
+
+_RAW_LEAK_MARKERS = ("ghp_", "password=", "api_key=", "ANTHROPIC_API_KEY")
+
+
+def _real_registry(workspace_root: Path) -> RunnerProfileRegistry:
+    yaml_path = (
+        Path(__file__).resolve().parents[3]
+        / "config"
+        / "workloads"
+        / "default-runner-profiles.yaml"
+    )
+    return RunnerProfileRegistry.load_file(yaml_path, workspace_root=workspace_root)
+
+
+def _assert_no_raw_blobs(result: dict) -> None:
+    """Result/workflow payload must not embed raw stdout/stderr/evidence bytes."""
+
+    serialized = repr(result)
+    for marker in _RAW_LEAK_MARKERS:
+        assert marker not in serialized, f"raw secret-like blob leaked: {marker}"
+
+
+@pytest.mark.skipif(
+    not (_E2E_ENABLED and _DOCKER_AVAILABLE),
+    reason="requires MOONMIND_PENTEST_E2E=1 and a Docker daemon",
+)
+async def test_recon_only_runs_through_real_docker_launcher(
+    tmp_path: Path, monkeypatch: pytest.MonkeyPatch
+) -> None:
+    monkeypatch.setattr(settings.pentest, "enabled", True)
+    workspace_root = tmp_path / "agent_jobs"
+    workspace_root.mkdir(parents=True, exist_ok=True)
+
+    from moonmind.workloads.docker_launcher import DockerWorkloadLauncher
+
+    launcher = DockerWorkloadLauncher(
+        docker_binary=os.environ.get("MOONMIND_DOCKER_BINARY", "docker"),
+        docker_host=os.environ.get("SYSTEM_DOCKER_HOST")
+        or os.environ.get("DOCKER_HOST"),
+    )
+    registry = _real_registry(workspace_root)
+
+    async with temporal_db(tmp_path) as session_maker:
+        async with session_maker() as session:
+            service = TemporalArtifactService(
+                TemporalArtifactRepository(session),
+                store=LocalTemporalArtifactStore(tmp_path / "artifact-store"),
+            )
+            activities = TemporalAgentRuntimeActivities(
+                artifact_service=service,
+                workload_launcher=launcher,
+                workload_registry=registry,
+            )
+
+            # Trusted-internal path with an inline lab scope keeps the test
+            # self-contained while still driving the real Docker launcher.
+            result = await activities._security_pentest_execute_trusted_internal(
+                _pentest_activity_payload(
+                    artifacts_dir=str(workspace_root / "runner-artifacts"),
+                    operation_mode="recon_only",
+                    approved_scope=_approved_pentest_scope(),
+                )
+            )
+
+    assert result["status"] in {"completed", "failed"}
+    # Report bundle is published using the shared report contract.
+    assert result.get("report_bundle_v") == 1
+    assert result.get("primary_report_ref")
+    assert result.get("structured_ref")
+    # Compact payload: counts present, no raw blobs embedded.
+    assert "findings_count" in result
+    _assert_no_raw_blobs(result)
+    # Terminal cleanup metadata is recorded for the container.
+    assert result.get("terminal_cleanup") or result.get("cleanup_metadata")
+
+
+@pytest.mark.skipif(
+    not (_E2E_ENABLED and _DOCKER_AVAILABLE),
+    reason="requires MOONMIND_PENTEST_E2E=1 and a Docker daemon",
+)
+async def test_invalid_scope_never_launches_docker(
+    tmp_path: Path, monkeypatch: pytest.MonkeyPatch
+) -> None:
+    """Untrusted dispatch with no artifact-backed scope must fail before Docker."""
+
+    monkeypatch.setattr(settings.pentest, "enabled", True)
+    workspace_root = tmp_path / "agent_jobs"
+    workspace_root.mkdir(parents=True, exist_ok=True)
+
+    from moonmind.workloads.docker_launcher import DockerWorkloadLauncher
+
+    launcher = DockerWorkloadLauncher()
+    registry = _real_registry(workspace_root)
+    activities = TemporalAgentRuntimeActivities(
+        workload_launcher=launcher,
+        workload_registry=registry,
+    )
+
+    result = await activities.security_pentest_execute(
+        _pentest_activity_payload(
+            scope_artifact_ref="",
+            approved_scope=_approved_pentest_scope(),
+            trusted_internal_execution=True,
+        )
+    )
+    assert result["status"] == "validation_failed"
+    assert "INVALID_SCOPE" in str(result["diagnostics"])
+
+
+@pytest.mark.skipif(
+    not _PROVIDER_E2E_ENABLED,
+    reason="requires MOONMIND_PENTEST_PROVIDER_E2E=1 and a provider profile manager",
+)
+async def test_provider_lease_lifecycle_against_real_manager(
+    tmp_path: Path, monkeypatch: pytest.MonkeyPatch
+) -> None:
+    """Phase 3: lease acquired only after scope validation and released on all paths.
+
+    This is a staging-only smoke test against the real provider profile manager
+    workflow. It is intentionally left as an opt-in skeleton: wire the real
+    ``pentest_provider_lease_manager`` and provider profile via environment, then
+    assert lease acquire/release ordering and redacted cooldown diagnostics.
+    """
+
+    pytest.skip(
+        "Provider-manager E2E must be run in staging with a configured "
+        "provider profile manager workflow and secret refs."
+    )
