+++ tests/unit/integrations/pentest/test_pentest_findings_normalization.py
@@ -1,0 +1,126 @@
+"""Phase 6: strict normalization of Pentest findings.
+
+These tests prove that runner/raw output is validated against the real
+``PentestFinding`` schema, malformed records are quarantined (never passed
+through as arbitrary dicts), the compact summary only counts valid findings,
+and the normalization status distinguishes "no vulnerabilities" from the
+several failure/empty modes so the report cannot imply a clean result on
+failure.
+"""
+
+from __future__ import annotations
+
+from datetime import UTC, datetime
+
+from moonmind.integrations.pentest import strictly_normalize_pentest_finding_set
+
+
+def _valid_finding(**overrides: object) -> dict[str, object]:
+    record = {
+        "finding_id": "f-1",
+        "title": "Auth bypass",
+        "severity": "high",
+        "confidence": "confirmed",
+        "target": "https://lab.example.test",
+        "summary": "Confirmed bypass with evidence.",
+    }
+    record.update(overrides)
+    return record
+
+
+def _normalize(payload, **kwargs):
+    kwargs.setdefault("target", "https://lab.example.test")
+    kwargs.setdefault("operation_mode", "validate_hypothesis")
+    kwargs.setdefault("runner_profile_id", "pentestgpt-safe")
+    kwargs.setdefault("produced_at", datetime(2026, 1, 1, tzinfo=UTC))
+    return strictly_normalize_pentest_finding_set(payload, **kwargs)
+
+
+def test_valid_findings_are_kept_and_counted() -> None:
+    finding_set, quarantined = _normalize({"findings": [_valid_finding()]})
+    assert quarantined == []
+    assert finding_set.normalization_status == "ok"
+    assert finding_set.summary.findings_count == 1
+    assert finding_set.summary.confirmed_findings_count == 1
+    assert finding_set.summary.high_or_critical_count == 1
+    assert finding_set.quarantined_findings_count == 0
+    assert not finding_set.implies_no_vulnerabilities
+
+
+def test_malformed_records_are_quarantined_not_passed_through() -> None:
+    finding_set, quarantined = _normalize(
+        {
+            "findings": [
+                _valid_finding(),
+                "not-an-object",
+                {"title": "missing required fields"},
+            ]
+        }
+    )
+    # Only the valid record survives into the schema-valid set.
+    assert finding_set.summary.findings_count == 1
+    assert finding_set.quarantined_findings_count == 2
+    assert {q["index"] for q in quarantined} == {1, 2}
+    # Quarantined records carry a redacted reason for restricted diagnostics.
+    assert all("reason" in q for q in quarantined)
+    assert finding_set.normalization_status == "ok"
+
+
+def test_all_invalid_records_yields_no_valid_findings_status() -> None:
+    finding_set, quarantined = _normalize(
+        {"findings": ["bad", {"nope": True}]}
+    )
+    assert finding_set.summary.findings_count == 0
+    assert finding_set.quarantined_findings_count == 2
+    assert finding_set.normalization_status == "no_valid_findings"
+    # A failure mode must never imply "no vulnerabilities".
+    assert not finding_set.implies_no_vulnerabilities
+
+
+def test_empty_findings_is_completed_no_findings() -> None:
+    finding_set, quarantined = _normalize({"findings": []})
+    assert quarantined == []
+    assert finding_set.normalization_status == "completed_no_findings"
+    assert finding_set.implies_no_vulnerabilities
+
+
+def test_missing_findings_key_is_no_machine_readable() -> None:
+    finding_set, _ = _normalize({"summary": {}})
+    assert finding_set.normalization_status == "no_machine_readable_findings"
+    assert not finding_set.implies_no_vulnerabilities
+
+
+def test_upstream_failure_is_not_reported_as_clean() -> None:
+    finding_set, _ = _normalize({"findings": []}, upstream_exit_code=21)
+    assert finding_set.normalization_status == "upstream_failed"
+    assert not finding_set.implies_no_vulnerabilities
+
+
+def test_explicit_status_hint_overrides() -> None:
+    finding_set, _ = _normalize({"findings": []}, status_hint="provider_failed")
+    assert finding_set.normalization_status == "provider_failed"
+    assert not finding_set.implies_no_vulnerabilities
+
+
+def test_none_payload_is_normalizer_error() -> None:
+    finding_set, _ = _normalize(None)
+    assert finding_set.normalization_status == "normalizer_error"
+    assert not finding_set.implies_no_vulnerabilities
+
+
+def test_target_backfilled_onto_findings_missing_target() -> None:
+    finding_set, quarantined = _normalize(
+        {"findings": [_valid_finding(target=None)]}
+    )
+    assert quarantined == []
+    assert finding_set.summary.findings_count == 1
+    assert finding_set.findings[0].target == "https://lab.example.test"
+
+
+def test_result_is_schema_valid_finding_set() -> None:
+    finding_set, _ = _normalize({"findings": [_valid_finding()]})
+    # Round-trips through the canonical schema without losing required fields.
+    dumped = finding_set.model_dump(mode="json")
+    assert dumped["tool_name"] == "security.pentest.run"
+    assert dumped["normalization_status"] == "ok"
+    assert "quarantined_findings_count" in dumped
