+++ tests/unit/integrations/pentest/test_pentest_models.py
@@ -491,7 +491,7 @@ def test_pentest_safe_profile_is_default_and_conservative() -> None:
     assert profile.kind == "one_shot"
     assert profile.image == "ghcr.io/moonladderstudios/moonmind-pentestgpt:1.0"
     assert profile.entrypoint == ("/usr/local/bin/moonmind-pentestgpt-run",)
-    assert profile.network_policy == "restricted_egress"
+    assert profile.network_policy == "bridge_approved_lab"
     assert profile.linux_capabilities == ()
     assert profile.devices == ()
     assert profile.cleanup == {
