+++ tests/unit/integrations/pentest/test_runner_normalize_findings.py
@@ -1,0 +1,87 @@
+"""Phase 5/6: in-runner findings normalizer (docker/pentestgpt-runner).
+
+The normalizer ships inside the curated runner image and cannot import
+MoonMind, so it is tested here by loading the script directly. It must mirror
+the strict schema contract: quarantine malformed records, count only valid
+findings, and classify the normalization status so the report never implies a
+clean result on failure.
+"""
+
+from __future__ import annotations
+
+import importlib.util
+from pathlib import Path
+
+_REPO_ROOT = Path(__file__).resolve().parents[4]
+_SCRIPT = _REPO_ROOT / "docker" / "pentestgpt-runner" / "normalize_findings.py"
+
+_spec = importlib.util.spec_from_file_location("_runner_normalize_findings", _SCRIPT)
+assert _spec and _spec.loader
+normalize_findings = importlib.util.module_from_spec(_spec)
+_spec.loader.exec_module(normalize_findings)
+
+
+def _norm(raw, *, target="https://lab.example.test", mode="recon_only"):
+    return normalize_findings.normalize_evidence(
+        raw, target=target, mode=mode, evidence_ref="file:/tmp/evidence.json"
+    )
+
+
+def _valid():
+    return {
+        "finding_id": "f-1",
+        "title": "Issue",
+        "severity": "high",
+        "confidence": "confirmed",
+        "target": "https://lab.example.test",
+        "summary": "Confirmed.",
+    }
+
+
+def test_valid_finding_counted_status_ok():
+    out = _norm({"findings": [_valid()]})
+    assert out["normalization_status"] == "ok"
+    assert out["summary"]["findings_count"] == 1
+    assert out["summary"]["high_or_critical_count"] == 1
+    assert out["quarantined_findings_count"] == 0
+    assert out["tool_name"] == "security.pentest.run"
+
+
+def test_malformed_records_quarantined():
+    out = _norm({"findings": [_valid(), "bad", {"title": "no id"}]})
+    assert out["summary"]["findings_count"] == 1
+    assert out["quarantined_findings_count"] == 2
+
+
+def test_all_invalid_is_no_valid_findings():
+    out = _norm({"findings": ["x", {"y": 1}]})
+    assert out["normalization_status"] == "no_valid_findings"
+    assert out["summary"]["findings_count"] == 0
+
+
+def test_empty_is_completed_no_findings():
+    out = _norm({"findings": []})
+    assert out["normalization_status"] == "completed_no_findings"
+
+
+def test_missing_findings_key_no_machine_readable():
+    out = _norm({"events": []})
+    assert out["normalization_status"] == "no_machine_readable_findings"
+
+
+def test_upstream_failure_status():
+    out = _norm({"findings": [], "upstream_exit_code": 21})
+    assert out["normalization_status"] == "upstream_failed"
+
+
+def test_unknown_confidence_downgraded_to_suspected():
+    out = _norm({"findings": [dict(_valid(), confidence="totally-sure")]})
+    assert out["findings"][0]["confidence"] == "suspected"
+
+
+def test_target_backfilled_when_missing():
+    record = _valid()
+    record.pop("target")
+    out = _norm({"findings": [record]})
+    assert out["summary"]["findings_count"] == 1
+    assert out["findings"][0]["target"] == "https://lab.example.test"
