+++ tests/unit/workflows/temporal/test_pentest_activity_extraction.py
@@ -1,0 +1,152 @@
+"""Phase 1 extraction guarantees for the curated PentestGPT activity.
+
+These tests prove that moving the Pentest-specific activity logic into the
+dedicated :class:`TemporalPentestActivities` class kept the public activity
+binding stable and preserved the trusted/untrusted authorization boundary:
+
+* ``_ACTIVITY_HANDLER_ATTRS["security.pentest.execute"]`` still resolves to the
+  public untrusted entrypoint on the agent-runtime activities;
+* trusted inline-scope execution remains internal-only (never registered);
+* untrusted execution can never self-authorize an inline ``approved_scope``.
+"""
+
+from __future__ import annotations
+
+import inspect
+
+import pytest
+
+from moonmind.config import settings
+from moonmind.workflows.temporal.activity_runtime import (
+    _ACTIVITY_HANDLER_ATTRS,
+    TemporalAgentRuntimeActivities,
+)
+from moonmind.workflows.temporal.activities.pentest_activities import (
+    TemporalPentestActivities,
+)
+from tests.unit.workflows.temporal.test_activity_runtime import (  # noqa: E501
+    _FakePentestLauncher,
+    _RecordingPentestRegistry,
+    _approved_pentest_scope,
+    _pentest_activity_payload,
+)
+
+
+def test_handler_attr_resolves_to_public_untrusted_entrypoint() -> None:
+    family, attr = _ACTIVITY_HANDLER_ATTRS["security.pentest.execute"]
+    assert family == "agent_runtime"
+    assert attr == "security_pentest_execute"
+
+    method = getattr(TemporalAgentRuntimeActivities, attr)
+    assert inspect.iscoroutinefunction(method)
+
+
+def test_trusted_internal_entrypoint_is_never_registered() -> None:
+    """The trusted inline-scope entrypoint must not be reachable via dispatch."""
+
+    registered_attrs = {attr for (_family, attr) in _ACTIVITY_HANDLER_ATTRS.values()}
+    assert "_security_pentest_execute_trusted_internal" not in registered_attrs
+
+
+def test_pentest_logic_lives_in_dedicated_class() -> None:
+    """The heavy Pentest logic moved off the generic runtime class."""
+
+    # The dedicated class owns the real orchestrator and helpers.
+    for name in (
+        "_run_security_pentest",
+        "_load_pentest_approved_scope",
+        "_validate_pentest_request_policy",
+        "_resolve_pentest_secret_env",
+        "_materialize_pentest_input_files",
+        "_publish_pentest_success_artifacts",
+    ):
+        assert hasattr(TemporalPentestActivities, name), name
+
+    # The generic runtime class keeps only thin delegates, not the orchestrator.
+    assert not hasattr(TemporalAgentRuntimeActivities, "_run_security_pentest")
+    assert not hasattr(TemporalAgentRuntimeActivities, "_load_pentest_approved_scope")
+
+
+def test_host_wires_dedicated_pentest_activities() -> None:
+    activities = TemporalAgentRuntimeActivities()
+    assert isinstance(activities._pentest_activities, TemporalPentestActivities)
+    # Shared collaborators resolve through the host (single source of truth).
+    assert (
+        activities._pentest_activities._artifact_service
+        is activities._artifact_service
+    )
+    assert (
+        activities._pentest_activities._pentest_provider_lease_manager
+        is activities._pentest_provider_lease_manager
+    )
+
+
+def test_public_entrypoint_delegates_to_dedicated_class(monkeypatch) -> None:
+    activities = TemporalAgentRuntimeActivities()
+    seen: dict[str, object] = {}
+
+    async def _fake(payload, /):
+        seen["payload"] = payload
+        return {"status": "delegated"}
+
+    monkeypatch.setattr(
+        activities._pentest_activities, "security_pentest_execute", _fake
+    )
+
+    import asyncio
+
+    result = asyncio.run(activities.security_pentest_execute({"probe": True}))
+    assert result == {"status": "delegated"}
+    assert seen["payload"] == {"probe": True}
+
+
+@pytest.mark.asyncio
+async def test_untrusted_public_entrypoint_rejects_inline_scope(monkeypatch) -> None:
+    """Untrusted dispatch must reject self-authorizing inline scope.
+
+    Even with a caller-supplied ``trusted_internal_execution`` flag, the public
+    entrypoint loads scope from the artifact store and never honors an inline
+    ``approved_scope`` — failing before any provider lease / docker launch.
+    """
+
+    monkeypatch.setattr(settings.pentest, "enabled", True)
+    launcher = _FakePentestLauncher()
+    activities = TemporalAgentRuntimeActivities(workload_launcher=launcher)
+
+    result = await activities.security_pentest_execute(
+        _pentest_activity_payload(
+            scope_artifact_ref=None,
+            approved_scope=_approved_pentest_scope(),
+            trusted_internal_execution=True,
+        )
+    )
+
+    assert result["status"] == "validation_failed"
+    assert "INVALID_SCOPE" in str(result["diagnostics"])
+    # No side effects: scope rejection happens before the docker launcher runs.
+    assert launcher.requests == []
+
+
+@pytest.mark.asyncio
+async def test_trusted_internal_entrypoint_allows_inline_scope_path(monkeypatch) -> None:
+    """The internal-only entrypoint honors an inline scope and reaches launch.
+
+    It must remain accessible only as a direct method call (never via dispatch),
+    which :func:`test_trusted_internal_entrypoint_is_never_registered` enforces.
+    """
+
+    monkeypatch.setattr(settings.pentest, "enabled", True)
+    activities = TemporalAgentRuntimeActivities(
+        workload_launcher=_FakePentestLauncher(),
+        workload_registry=_RecordingPentestRegistry(),
+    )
+
+    result = await activities._security_pentest_execute_trusted_internal(
+        _pentest_activity_payload(
+            scope_artifact_ref=None,
+            trusted_internal_execution=True,
+        )
+    )
+
+    assert result["status"] == "completed"
+    assert result["launch_plan"]["profile_id"] == "pentestgpt-safe"
